Group users routes
Ref: #550
