+++ public/js/app.js
@@ -1,1 +1,1 @@
-(window.webpackJsonp=window.webpackJsonp||[]).push([[0],{"+jwq":function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"#main-navigation .router-link-active[data-v-01da79fe] {\n  color: #3490dc;\n}",""])},"/BTM":function(t,e,n){var s=n("fysm");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},0:function(t,e,n){n("bUC5"),t.exports=n("pyCd")},"0P3G":function(t,e,n){"use strict";var s=n("/BTM");n.n(s).a},"6pLF":function(t,e,n){var s=n("shRJ");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},"7HpC":function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\ntextarea[data-v-4b83dcc8] {\n  outline: 0;\n  background: #f4f4f4;\n  box-shadow: 1px 1px 0 #ccc;\n  border-color: transparent;\n  border-radius: 0;\n  font-family: 'Inconsolata', monospace;\n}\ntextarea[data-v-4b83dcc8]:focus {\n  box-shadow: 1px 1px 0 #ccc;\n  outline: 0;\n  background: #f4f4f4;\n  border-color: transparent;\n}\n.new-tag-group[data-v-4b83dcc8] {\n  font-size: 0;\n}\n.new-tag-group .input-new-tag[data-v-4b83dcc8], .new-tag-group .button-new-tag[data-v-4b83dcc8] {\n  vertical-align: middle;\n  width: 120px;\n}\n#note-update[data-v-4b83dcc8] .el-checkbox-button--small .el-checkbox-button__inner {\n  border-radius: 4px;\n  border: 1px solid #DCDFE6;\n}\n#note-update[data-v-4b83dcc8] .el-checkbox-button.is-checked .el-checkbox-button__inner {\n  border-color: #409EFF;\n}\n",""])},"7Pbq":function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,'\n.loader-text[data-v-1e7e8ff5] {\n  background: #808080;\n  border-radius: 2px;\n}\n.loader-text--1x3[data-v-1e7e8ff5] {\n  width: 33%;\n}\n.note[data-v-1e7e8ff5] .footnotes {\n  font-size: 0.8rem;\n}\n.note[data-v-1e7e8ff5] .footnotes p {\n  margin-bottom: 0.5rem;\n}\n.note[data-v-1e7e8ff5] table.table th[align="left"] {\n  text-align: left;\n}\n.note[data-v-1e7e8ff5] table.table th[align="center"] {\n  text-align: center;\n}\n.note[data-v-1e7e8ff5] table.table th[align="right"] {\n  text-align: right;\n}\n',""])},Boyg:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.form-control[data-v-396d57c6] {\n  width: auto;\n}\n",""])},CbGa:function(t,e,n){var s=n("HPB2");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},"D1a/":function(t,e,n){var s=n("WGSt");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},EPOQ:function(t,e,n){"use strict";var s=n("veFt");n.n(s).a},F4tQ:function(t,e,n){"use strict";var s=n("WwcV");n.n(s).a},GYTP:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.slide-fade-enter-active[data-v-764fdbfc] {\n  transition: all .3s ease;\n}\n.slide-fade-leave-active[data-v-764fdbfc] {\n  transition: all .5s cubic-bezier(1.0, 0.5, 0.8, 1.0);\n}\n.slide-fade-enter[data-v-764fdbfc], .slide-fade-leave-to[data-v-764fdbfc] {\n  background-color: #ffef0029;\n}\n",""])},HPB2:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.divider[data-v-01380892] {\n  border-left: 1px solid #ccc;\n}\n",""])},"I/cB":function(t,e,n){"use strict";var s=n("aS7X");n.n(s).a},J82J:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.slide-enter-active[data-v-1a92697e] {\n  transition-duration: 0.3s;\n  transition-timing-function: ease-in;\n}\n.slide-leave-active[data-v-1a92697e] {\n  transition-duration: 0.3s;\n  transition-timing-function: cubic-bezier(0, 1, 0.5, 1);\n}\n.slide-enter-to[data-v-1a92697e], .slide-leave[data-v-1a92697e] {\n  max-height: 1000px;\n  overflow: hidden;\n}\n.slide-enter[data-v-1a92697e], .slide-leave-to[data-v-1a92697e] {\n  overflow: hidden;\n  max-height: 0;\n}\n.text-line-through[data-v-1a92697e] {\n  text-decoration: line-through;\n}\n.btn-collapse[data-v-1a92697e] {\n  font-family: 'Inconsolata', monospace;\n  padding: 0;\n  height: 23px;\n}\n",""])},JRqE:function(t,e,n){var s=n("7HpC");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},JVNZ:function(t,e,n){"use strict";var s=n("6pLF");n.n(s).a},L7bX:function(t,e,n){"use strict";var s=n("CbGa");n.n(s).a},"O//m":function(t,e,n){"use strict";var s=n("JRqE");n.n(s).a},"Od/V":function(t,e,n){"use strict";var s=n("hW3y");n.n(s).a},P9EP:function(t,e,n){"use strict";var s=n("avtZ");n.n(s).a},RnhZ:function(t,e,n){var s={"./af":"K/tc","./af.js":"K/tc","./ar":"jnO4","./ar-dz":"o1bE","./ar-dz.js":"o1bE","./ar-kw":"Qj4J","./ar-kw.js":"Qj4J","./ar-ly":"HP3h","./ar-ly.js":"HP3h","./ar-ma":"CoRJ","./ar-ma.js":"CoRJ","./ar-sa":"gjCT","./ar-sa.js":"gjCT","./ar-tn":"bYM6","./ar-tn.js":"bYM6","./ar.js":"jnO4","./az":"SFxW","./az.js":"SFxW","./be":"H8ED","./be.js":"H8ED","./bg":"hKrs","./bg.js":"hKrs","./bm":"p/rL","./bm.js":"p/rL","./bn":"kEOa","./bn.js":"kEOa","./bo":"0mo+","./bo.js":"0mo+","./br":"aIdf","./br.js":"aIdf","./bs":"JVSJ","./bs.js":"JVSJ","./ca":"1xZ4","./ca.js":"1xZ4","./cs":"PA2r","./cs.js":"PA2r","./cv":"A+xa","./cv.js":"A+xa","./cy":"l5ep","./cy.js":"l5ep","./da":"DxQv","./da.js":"DxQv","./de":"tGlX","./de-at":"s+uk","./de-at.js":"s+uk","./de-ch":"u3GI","./de-ch.js":"u3GI","./de.js":"tGlX","./dv":"WYrj","./dv.js":"WYrj","./el":"jUeY","./el.js":"jUeY","./en-au":"Dmvi","./en-au.js":"Dmvi","./en-ca":"OIYi","./en-ca.js":"OIYi","./en-gb":"Oaa7","./en-gb.js":"Oaa7","./en-ie":"4dOw","./en-ie.js":"4dOw","./en-il":"czMo","./en-il.js":"czMo","./en-in":"7C5Q","./en-in.js":"7C5Q","./en-nz":"b1Dy","./en-nz.js":"b1Dy","./en-sg":"t+mt","./en-sg.js":"t+mt","./eo":"Zduo","./eo.js":"Zduo","./es":"iYuL","./es-do":"CjzT","./es-do.js":"CjzT","./es-us":"Vclq","./es-us.js":"Vclq","./es.js":"iYuL","./et":"7BjC","./et.js":"7BjC","./eu":"D/JM","./eu.js":"D/JM","./fa":"jfSC","./fa.js":"jfSC","./fi":"gekB","./fi.js":"gekB","./fil":"1ppg","./fil.js":"1ppg","./fo":"ByF4","./fo.js":"ByF4","./fr":"nyYc","./fr-ca":"2fjn","./fr-ca.js":"2fjn","./fr-ch":"Dkky","./fr-ch.js":"Dkky","./fr.js":"nyYc","./fy":"cRix","./fy.js":"cRix","./ga":"USCx","./ga.js":"USCx","./gd":"9rRi","./gd.js":"9rRi","./gl":"iEDd","./gl.js":"iEDd","./gom-deva":"qvJo","./gom-deva.js":"qvJo","./gom-latn":"DKr+","./gom-latn.js":"DKr+","./gu":"4MV3","./gu.js":"4MV3","./he":"x6pH","./he.js":"x6pH","./hi":"3E1r","./hi.js":"3E1r","./hr":"S6ln","./hr.js":"S6ln","./hu":"WxRl","./hu.js":"WxRl","./hy-am":"1rYy","./hy-am.js":"1rYy","./id":"UDhR","./id.js":"UDhR","./is":"BVg3","./is.js":"BVg3","./it":"bpih","./it-ch":"bxKX","./it-ch.js":"bxKX","./it.js":"bpih","./ja":"B55N","./ja.js":"B55N","./jv":"tUCv","./jv.js":"tUCv","./ka":"IBtZ","./ka.js":"IBtZ","./kk":"bXm7","./kk.js":"bXm7","./km":"6B0Y","./km.js":"6B0Y","./kn":"PpIw","./kn.js":"PpIw","./ko":"Ivi+","./ko.js":"Ivi+","./ku":"JCF/","./ku.js":"JCF/","./ky":"lgnt","./ky.js":"lgnt","./lb":"RAwQ","./lb.js":"RAwQ","./lo":"sp3z","./lo.js":"sp3z","./lt":"JvlW","./lt.js":"JvlW","./lv":"uXwI","./lv.js":"uXwI","./me":"KTz0","./me.js":"KTz0","./mi":"aIsn","./mi.js":"aIsn","./mk":"aQkU","./mk.js":"aQkU","./ml":"AvvY","./ml.js":"AvvY","./mn":"lYtQ","./mn.js":"lYtQ","./mr":"Ob0Z","./mr.js":"Ob0Z","./ms":"6+QB","./ms-my":"ZAMP","./ms-my.js":"ZAMP","./ms.js":"6+QB","./mt":"G0Uy","./mt.js":"G0Uy","./my":"honF","./my.js":"honF","./nb":"bOMt","./nb.js":"bOMt","./ne":"OjkT","./ne.js":"OjkT","./nl":"+s0g","./nl-be":"2ykv","./nl-be.js":"2ykv","./nl.js":"+s0g","./nn":"uEye","./nn.js":"uEye","./oc-lnc":"Fnuy","./oc-lnc.js":"Fnuy","./pa-in":"8/+R","./pa-in.js":"8/+R","./pl":"jVdC","./pl.js":"jVdC","./pt":"8mBD","./pt-br":"0tRk","./pt-br.js":"0tRk","./pt.js":"8mBD","./ro":"lyxo","./ro.js":"lyxo","./ru":"lXzo","./ru.js":"lXzo","./sd":"Z4QM","./sd.js":"Z4QM","./se":"//9w","./se.js":"//9w","./si":"7aV9","./si.js":"7aV9","./sk":"e+ae","./sk.js":"e+ae","./sl":"gVVK","./sl.js":"gVVK","./sq":"yPMs","./sq.js":"yPMs","./sr":"zx6S","./sr-cyrl":"E+lV","./sr-cyrl.js":"E+lV","./sr.js":"zx6S","./ss":"Ur1D","./ss.js":"Ur1D","./sv":"X709","./sv.js":"X709","./sw":"dNwA","./sw.js":"dNwA","./ta":"PeUW","./ta.js":"PeUW","./te":"XLvN","./te.js":"XLvN","./tet":"V2x9","./tet.js":"V2x9","./tg":"Oxv6","./tg.js":"Oxv6","./th":"EOgW","./th.js":"EOgW","./tl-ph":"Dzi0","./tl-ph.js":"Dzi0","./tlh":"z3Vd","./tlh.js":"z3Vd","./tr":"DoHr","./tr.js":"DoHr","./tzl":"z1FC","./tzl.js":"z1FC","./tzm":"wQk9","./tzm-latn":"tT3J","./tzm-latn.js":"tT3J","./tzm.js":"wQk9","./ug-cn":"YRex","./ug-cn.js":"YRex","./uk":"raLr","./uk.js":"raLr","./ur":"UpQW","./ur.js":"UpQW","./uz":"Loxo","./uz-latn":"AQ68","./uz-latn.js":"AQ68","./uz.js":"Loxo","./vi":"KSF8","./vi.js":"KSF8","./x-pseudo":"/X5v","./x-pseudo.js":"/X5v","./yo":"fzPg","./yo.js":"fzPg","./zh-cn":"XDpg","./zh-cn.js":"XDpg","./zh-hk":"SatO","./zh-hk.js":"SatO","./zh-mo":"OmwH","./zh-mo.js":"OmwH","./zh-tw":"kOpN","./zh-tw.js":"kOpN"};function a(t){var e=r(t);return n(e)}function r(t){if(!n.o(s,t)){var e=new Error("Cannot find module '"+t+"'");throw e.code="MODULE_NOT_FOUND",e}return s[t]}a.keys=function(){return Object.keys(s)},a.resolve=r,t.exports=a,a.id="RnhZ"},WGSt:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.bookmark-story[data-v-cdd94d26] {\n  position: relative;\n  top: 2px;\n  text-decoration: none;\n}\n.loader-text[data-v-cdd94d26] {\n  background: #808080;\n  border-radius: 2px;\n}\n.loader-text--1x3[data-v-cdd94d26] {\n  width: 33%;\n}\n.loader-text--2x3[data-v-cdd94d26] {\n  width: 66%;\n}\n",""])},WTCb:function(t,e,n){"use strict";var s=n("k1kw");n.n(s).a},WobJ:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\ntextarea[data-v-c5e2bf76] {\n  outline: 0;\n  background: #f4f4f4;\n  box-shadow: 1px 1px 0 #ccc;\n  border-color: transparent;\n  border-radius: 0;\n  font-family: 'Inconsolata', monospace;\n}\ntextarea[data-v-c5e2bf76]:focus {\n  box-shadow: 1px 1px 0 #ccc;\n  outline: 0;\n  background: #f4f4f4;\n  border-color: transparent;\n}\n.new-tag-group[data-v-c5e2bf76] {\n  font-size: 0;\n}\n.new-tag-group .input-new-tag[data-v-c5e2bf76], .new-tag-group .button-new-tag[data-v-c5e2bf76] {\n  vertical-align: middle;\n  width: 120px;\n}\n#note-create[data-v-c5e2bf76] .el-checkbox-button--small .el-checkbox-button__inner {\n  border-radius: 4px;\n  border: 1px solid #DCDFE6;\n}\n#note-create[data-v-c5e2bf76] .el-checkbox-button.is-checked .el-checkbox-button__inner {\n  border-color: #409EFF;\n}\n",""])},WwcV:function(t,e,n){var s=n("GYTP");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},a5qU:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n#users[data-v-051ebd10] .el-table .success-row {\n  background: #f0f9eb;\n}\n@media (max-width: 575px) {\n#users[data-v-051ebd10] .el-dialog {\n    width: 100%;\n}\n}\n",""])},aS7X:function(t,e,n){var s=n("+jwq");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},avtZ:function(t,e,n){var s=n("J82J");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},bUC5:function(t,e,n){"use strict";n.r(e);var s=n("XuX8"),a=n.n(s),r=n("jE9Z"),o=n("o0o1"),i=n.n(o),c=n("yDJ3"),l=n.n(c),u=n("wd/R"),d=n.n(u);function m(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var p={name:"Weather",data:function(){return{loading:!0,weather:{icon:"",tempFeelsLike:"",humidity:"",description:"",sunrise:"",sunset:""}}},methods:{fetchWeather:function(){var t,e=this;return(t=i.a.mark((function t(){var n,s,a,r,o;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:n="openweather--currentweather",a=localStorage.getItem(n);try{s=JSON.parse(a)}catch(t){s=null}if(s&&s.data&&!(s.saved+6e5<Date.now())){t.next=12;break}return s={data:{},saved:Date.now()},t.next=7,fetch("http://api.openweathermap.org/data/2.5/weather?q=Funchal,PT&appid=fe66d26d5b01501290be42a4809bbc4e&units=metric&lang=pt");case 7:return r=t.sent,t.next=10,r.json();case 10:s.data=t.sent,localStorage.setItem(n,JSON.stringify(s));case 12:o=l()(s,"data.weather[0].icon"),e.weather.icon="http://openweathermap.org/img/wn/".concat(o,".png"),e.weather.tempFeelsLike=l()(s,"data.main.feels_like"),e.weather.temp=l()(s,"data.main.temp"),e.weather.description=l()(s,"data.weather[0].description"),e.weather.humidity=l()(s,"data.main.humidity"),e.weather.sunrise=l()(s,"data.sys.sunrise"),e.weather.sunset=l()(s,"data.sys.sunset"),setTimeout((function(){return e.loading=!1}),500);case 21:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){m(r,s,a,o,i,"next",t)}function i(t){m(r,s,a,o,i,"throw",t)}o(void 0)}))})()}},filters:{capitalize:function(t){if(t){var e=t.toString();return e.charAt(0).toUpperCase()+e.slice(1)}},localTimeFromUnixTimestamp:function(t){if(t)return d.a.unix(t).format("HH:mm")}},created:function(){this.fetchWeather(),setInterval(this.fetchWeather,6e5)}},f=(n("L7bX"),n("KHd+")),v=Object(f.a)(p,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],attrs:{id:"weather"}},[n("div",{staticClass:"card mb-3 shadow"},[n("div",{staticClass:"card-body d-flex flex-wrap"},[n("div",{staticClass:"align-self-center"},[n("a",{attrs:{href:"https://openweathermap.org/city/2267827",target:"_blank"}},[n("img",{attrs:{src:t.weather.icon,alt:t.weather.description}})])]),t._v(" "),n("div",{staticClass:"divider align-items-stretch mx-2"}),t._v(" "),n("div",{staticClass:"align-self-center ml-2"},[t._v("\n        "+t._s(t._f("capitalize")(t.weather.description))+"\n      ")]),t._v(" "),n("div",{staticClass:"divider align-items-stretch mx-2"}),t._v(" "),n("div",{staticClass:"align-self-center"},[n("div",{staticClass:"d-flex"},[n("div",{staticClass:"align-items-stretch align-self-center"},[t._v("Temps:")]),t._v(" "),n("div",{staticClass:"d-flex flex-column ml-1"},[n("div",{attrs:{title:"Feels like"}},[t._v(t._s(t.weather.tempFeelsLike)+" ℃")]),t._v(" "),n("div",{attrs:{title:"Temperature"}},[n("small",[t._v(t._s(t.weather.temp)+" ℃")])])])])]),t._v(" "),n("div",{staticClass:"divider align-items-stretch mx-2"}),t._v(" "),n("div",{staticClass:"align-self-center"},[n("div",{staticClass:"d-flex"},[n("div",{staticClass:"flex-column"},[n("div",[t._v("Humidade")]),t._v(" "),n("div",{staticClass:"text-center"},[t._v(t._s(t.weather.humidity)+"%")])])])]),t._v(" "),n("div",{staticClass:"divider align-items-stretch mx-2"}),t._v(" "),n("div",{staticClass:"align-self-center"},[n("div",{staticClass:"d-flex"},[n("div",{staticClass:"flex-column ml-1"},[n("div",{attrs:{title:"Sunrise"}},[t._v(t._s(t._f("localTimeFromUnixTimestamp")(t.weather.sunrise)))]),t._v(" "),n("div",{attrs:{title:"Sunset"}},[t._v(t._s(t._f("localTimeFromUnixTimestamp")(t.weather.sunset)))])])])])])])])}),[],!1,null,"01380892",null).exports,h=n("vDqi"),g=n.n(h);function b(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var _={name:"SystemInfo",data:function(){return{loading:!0,memory:{free:"",used:"",total:""},nQueueWorkersRunning:0,up:"",upSince:"",systemInfoLoop:null}},created:function(){this.fetchSystemInfo(),this.systemInfoLoop=setInterval(this.fetchSystemInfo,1e4)},methods:{fetchSystemInfo:function(){var t,e=this;return(t=i.a.mark((function t(){var n,s;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,g.a.get("/api/system-info");case 2:n=t.sent,(s=l()(n,"data.data",null))&&(e.memory.free=l()(s,"memory.free"),e.memory.used=l()(s,"memory.used"),e.memory.total=l()(s,"memory.total"),e.nQueueWorkersRunning=l()(s,"nQueueWorkersRunning"),e.up=l()(s,"up"),e.upSince=l()(s,"upSince")),setTimeout((function(){return e.loading=!1}),500);case 6:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){b(r,s,a,o,i,"next",t)}function i(t){b(r,s,a,o,i,"throw",t)}o(void 0)}))})()}},filters:{capitalize:function(t){if(t)return(t=t.toString()).charAt(0).toUpperCase()+t.slice(1)}}},x=(n("F4tQ"),Object(f.a)(_,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"card mb-3 shadow"},[n("div",{staticClass:"card-header"},[t._v(t._s(t.$I18n.trans("system.info")))]),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"card-body"},[n("transition",{attrs:{name:"slide-fade",mode:"out-in"}},[n("p",{key:t.up,staticClass:"alert-info"},[t._v("\n        "+t._s(t._f("capitalize")(t.up))+" "+t._s(t.$I18n.trans("system.since"))+" "+t._s(t.upSince)+"\n      ")])]),t._v(" "),n("p",{class:{"alert-success":t.nQueueWorkersRunning,"alert-danger":!t.nQueueWorkersRunning}},[t._v("\n      "+t._s(t.$I18n.trans("system.number_queue_workers"))+": "+t._s(t.nQueueWorkersRunning)+"\n    ")]),t._v(" "),n("div",{staticClass:"pt-4"},[n("table",{staticClass:"table table-sm"},[n("caption",[t._v(t._s(t.$I18n.trans("system.since")))]),t._v(" "),n("thead",[n("tr",[n("th",[t._v(t._s(t.$I18n.trans("system.used")))]),t._v(" "),n("th",[t._v(t._s(t.$I18n.trans("system.free")))]),t._v(" "),n("th",[t._v(t._s(t.$I18n.trans("system.total")))])])]),t._v(" "),n("tbody",[n("transition",{attrs:{name:"slide-fade",mode:"out-in"}},[n("tr",{key:t.memory.used},[n("td",[t.memory.used?t._e():n("span",[t._v(" ")]),t._v(t._s(t.memory.used))]),t._v(" "),n("td",[t._v(t._s(t.memory.free))]),t._v(" "),n("td",[t._v(t._s(t.memory.total))])])])],1)])])],1)])}),[],!1,null,"764fdbfc",null).exports);function k(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var y={name:"Inspire",data:function(){return{inspire:null,fallbackLines:["The hardest part of building software is deciding what to build.","Everything not specified will be specified by the programmer, on the fly.","Design is also specification.","Everything that is public-facing defines the significant part of the business logic.","The open office plan is the best way to destroy productivity.","Don’t multitask. Focus on one task at a time.","There is work time and there is personal time. Don’t mix the two.","Everyone writes bad code except for you. /s"]}},computed:{fallback:function(){return this.fallbackLines[Math.floor(Math.random()*this.fallbackLines.length)]}},created:function(){this.fetchInspire()},methods:{fetchInspire:function(){var t,e=this;return(t=i.a.mark((function t(){var n;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,g.a.get("/api/inspire");case 2:n=t.sent,e.inspire=l()(n,"data.data");case 4:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){k(r,s,a,o,i,"next",t)}function i(t){k(r,s,a,o,i,"throw",t)}o(void 0)}))})()}}},w=Object(f.a)(y,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"card mb-3 shadow"},[n("div",{staticClass:"card-body"},[n("small",[t.inspire?t._e():n("span",[t._v(t._s(t.fallback))]),t._v(t._s(t.inspire))])])])}),[],!1,null,null,null).exports;function C(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var j={name:"NextHolidays",data:function(){return{loading:!0,nextHolidays:[{name:"",date:"",type:"",description:""},{name:"",date:"",type:"",description:""},{name:"",date:"",type:"",description:""}]}},methods:{fetchNextHolidays:function(){var t,e=this;return(t=i.a.mark((function t(){var n,s;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,g.a.get("/api/next-holidays");case 2:return n=t.sent,(s=l()(n,"data.data",[])).length&&(e.nextHolidays=s.map((function(t){return{name:t.name,description:t.description,date:t.date.iso,type:Array.isArray(t.type)?t.type[0]:""}})),e.loading=!1),t.abrupt("return",!0);case 6:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){C(r,s,a,o,i,"next",t)}function i(t){C(r,s,a,o,i,"throw",t)}o(void 0)}))})()}},created:function(){this.fetchNextHolidays()}},I={name:"Dashboard",components:{Weather:v,SystemInfo:x,NextHolidays:Object(f.a)(j,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"card shadow mb-3"},[n("div",{staticClass:"card-header"},[t._v(t._s(t.$I18n.trans("holidays.next_holidays")))]),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"card-body"},t._l(t.nextHolidays,(function(e){return n("div",[n("p",[n("span",{staticClass:"badge badge-dark"},[t._v("#")]),t._v("\n        "+t._s(e.name)+",\n        "+t._s(t._f("diffForHumans")(e.date))+"\n        at "+t._s(e.date)+"\n        "),n("small",[t._v(t._s(e.type))])]),t._v(" "),n("p",[e.description?t._e():n("span",[t._v(" ")]),t._v(" "),n("small",[t._v(t._s(e.description))])])])})),0)])}),[],!1,null,null,null).exports,Inspire:w},data:function(){return{}},beforeRouteLeave:function(t,e,n){clearInterval(this.$refs.refSystemInfo.systemInfoLoop),n()}},$=Object(f.a)(I,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{staticClass:"pt-4",attrs:{id:"view-dashboard"}},[e("div",{staticClass:"row justify-content-center"},[e("div",{staticClass:"col-sm-6 no-gutter-xs"},[e("inspire"),this._v(" "),e("system-info",{ref:"refSystemInfo"})],1),this._v(" "),e("div",{staticClass:"col-sm-6 no-gutter-xs"},[e("weather"),this._v(" "),e("next-holidays")],1)])])}),[],!1,null,null,null).exports;function T(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}function E(t){return function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){T(r,s,a,o,i,"next",t)}function i(t){T(r,s,a,o,i,"throw",t)}o(void 0)}))}}var S={name:"Authenticate",data:function(){return{user:{email:"",password:""},pollingPasswordFieldInterval:null,errorMessage:null}},methods:{pollingPasswordField:function(){var t=this;this.pollingPasswordFieldInterval=setInterval((function(){var e=document.createEvent("HTMLEvents");e.initEvent("change",!1,!0),t.$refs.pwf.dispatchEvent(e)}),500)},passwordChanged:function(){this.user.password=this.$refs.pwf.value},attemptLogin:function(t){var e=this;return E(i.a.mark((function n(){return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return t.preventDefault(),n.next=3,g.a.post("/logout");case 3:return n.next=5,g.a.get("/sanctum/csrf-cookie");case 5:return n.next=7,e.login();case 7:case"end":return n.stop()}}),n)})))()},login:function(){var t=this;return E(i.a.mark((function e(){var n,s,a,r,o;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n=422,s=302,a={validateStatus:function(t){return t>=200&&t<300||t===n||t===s}},e.next=5,g.a.post("/login",t.user,a);case 5:if(r=e.sent,(o=r.status)!==n){e.next=11;break}t.showsErrors(l()(r,"data.errors")),e.next=16;break;case 11:if(o!==s&&204!==o){e.next=16;break}return clearInterval(t.pollingPasswordFieldInterval),t.$message.success("Login successful"),e.next=16,t.$router.push(Pt);case 16:case"end":return e.stop()}}),e)})))()},showsErrors:function(t){var e="";Object.keys(t).forEach((function(n){t[n].forEach((function(t){e+=t+"<br/>"}))})),this.$message({type:"error",dangerouslyUseHTMLString:!0,message:e})}},mounted:function(){this.pollingPasswordField()}},N=Object(f.a)(S,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"pt-4",attrs:{id:"view-authenticate"}},[n("div",{staticClass:"row justify-content-center"},[n("div",{staticClass:"col-md-8"},[n("div",{staticClass:"card shadow"},[n("div",{staticClass:"card-header"},[t._v(t._s(t.$I18n.trans("auth.login")))]),t._v(" "),n("div",{staticClass:"card-body"},[n("form",{attrs:{method:"POST",model:t.user},on:{submit:function(e){return t.attemptLogin(e)}}},[n("div",{staticClass:"form-group row"},[n("label",{staticClass:"col-md-4 col-form-label text-md-right",attrs:{for:"email"}},[t._v(t._s(t.$I18n.trans("auth.email")))]),t._v(" "),n("div",{staticClass:"col-md-6"},[n("input",{directives:[{name:"model",rawName:"v-model",value:t.user.email,expression:"user.email"}],staticClass:"form-control",attrs:{id:"email",inputmode:"email",type:"email",name:"email",autocomplete:"email",required:""},domProps:{value:t.user.email},on:{input:function(e){e.target.composing||t.$set(t.user,"email",e.target.value)}}})])]),t._v(" "),n("div",{staticClass:"form-group row"},[n("label",{staticClass:"col-md-4 col-form-label text-md-right",attrs:{for:"password"}},[t._v(t._s(t.$I18n.trans("auth.password")))]),t._v(" "),n("div",{staticClass:"col-md-6"},[n("input",{directives:[{name:"model",rawName:"v-model",value:t.user.password,expression:"user.password"}],ref:"pwf",staticClass:"form-control",attrs:{id:"password",name:"password",type:"password",autocomplete:"current-password",required:""},domProps:{value:t.user.password},on:{change:t.passwordChanged,input:function(e){e.target.composing||t.$set(t.user,"password",e.target.value)}}})])]),t._v(" "),n("div",{staticClass:"form-group row"},[n("div",{staticClass:"col-md-6 offset-md-4"},[n("div",{staticClass:"form-check"},[n("input",{staticClass:"form-check-input",attrs:{type:"checkbox",name:"remember",id:"remember",checked:""}}),t._v(" "),n("label",{staticClass:"form-check-label",attrs:{for:"remember"}},[t._v(t._s(t.$I18n.trans("auth.remember_me")))])])])]),t._v(" "),n("div",{staticClass:"form-group row mb-0"},[n("div",{staticClass:"col-md-8 offset-md-4"},[n("button",{staticClass:"btn btn-primary",attrs:{type:"submit",disabled:!t.user.email&&!t.user.password}},[t._v(t._s(t.$I18n.trans("auth.login")))])])])])])])])])])}),[],!1,null,null,null).exports,P={name:"Navigation",data:function(){return{notesRoute:Ot,notesCreateRoute:Bt}},methods:{activeSubmenu:function(t){return this.$route.name===t?"btn-primary":"btn-secondary"}}},O=Object(f.a)(P,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100",class:t.activeSubmenu("Notes"),attrs:{to:t.notesRoute}},[t._v(t._s(t.$I18n.trans("notes.notes")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100",class:t.activeSubmenu("NotesCreate"),attrs:{to:t.notesCreateRoute}},[t._v(t._s(t.$I18n.trans("notes.new_note")))])],1)])])])}),[],!1,null,null,null).exports,B=n("9/5/"),D={name:"AlgoliaSearch",data:function(){return{searchQuery:""}},watch:{searchQuery:function(){(0===this.searchQuery.length||this.searchQuery.length>1)&&this.submit()}},methods:{submit:n.n(B)()((function(){this.$emit("inputData",this.searchQuery)}),400),clearSearch:function(){this.searchQuery="",this.submit()}}};n("maHV");function U(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var R={name:"NotesList",props:["searchQuery"],data:function(){return{notesShowRoute:Dt,loading:!0,notes:[]}},created:function(){var t=this;this.fetchNotes(this.searchQuery).then((function(e){t.loading=!1}))},watch:{searchQuery:function(){var t=this;this.loading=!0,this.fetchNotes(this.searchQuery).then((function(e){t.loading=!1}))}},methods:{fetchNotes:function(t){var e,n=this;return(e=i.a.mark((function e(){var s;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,g.a.get("/api/notes?query=".concat(t));case 2:return s=e.sent,n.notes=l()(s,"data.data",[]),e.abrupt("return",!0);case 5:case"end":return e.stop()}}),e)})),function(){var t=this,n=arguments;return new Promise((function(s,a){var r=e.apply(t,n);function o(t){U(r,s,a,o,i,"next",t)}function i(t){U(r,s,a,o,i,"throw",t)}o(void 0)}))})()},close:function(){this.$emit("inputData","")}}},F={name:"Notes",data:function(){return{searchQuery:""}},components:{Navigation:O,AlgoliaSearch:Object(f.a)(D,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"col-12 no-gutter-xs",attrs:{id:"algolia-search"}},[n("div",{staticClass:"justify-content-center form-inline d-flex"},[n("label",{staticClass:"sr-only",attrs:{for:"query"}},[t._v(t._s(t.$I18n.trans("notes.search")))]),t._v(" "),n("input",{directives:[{name:"model",rawName:"v-model",value:t.searchQuery,expression:"searchQuery"}],staticClass:"form-control flex-grow-1",attrs:{type:"text",id:"query",name:"query",placeholder:t.$I18n.trans("notes.min_2_chars")},domProps:{value:t.searchQuery},on:{keyup:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"esc",27,e.key,["Esc","Escape"])?null:t.clearSearch(e)},input:function(e){e.target.composing||(t.searchQuery=e.target.value)}}}),t._v("\n     \n    "),t._m(0)])])}),[function(){var t=this.$createElement,e=this._self._c||t;return e("picture",[e("source",{attrs:{srcset:"search-by-algolia.svg",media:"(min-width: 800px)"}}),this._v(" "),e("img",{staticStyle:{"max-height":"30px"},attrs:{src:"algolia-blue-mark.svg",alt:"search by algolia"}})])}],!1,null,"396d57c6",null).exports,NotesList:Object(f.a)(R,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"notes-list"}},[n("ul",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"list-group list-group-flush"},[t.notes.length?t._e():n("el-alert",{attrs:{title:t.$I18n.trans("notes.alert_failed_to_find_records"),type:"info",center:"","close-text":t.$I18n.trans("notes.alert_close_text")},on:{close:t.close}}),t._v(" "),t._l(t.notes,(function(e){return n("li",{staticClass:"list-group-item list-group-item-action p-2 p-sm-3"},[n("router-link",{staticClass:"text-secondary",attrs:{to:{name:t.notesShowRoute.name,params:{id:e.id}}}},[n("div",{staticClass:"d-flex w-100 justify-content-between"},[n("h5",{staticClass:"mb-1"},[t._v(t._s(e.title))]),t._v(" "),n("small",{staticClass:"d-none d-sm-block"},[t._v(t._s(t._f("diffForHumans")(e.updated_at)))])])]),t._v(" "),n("div",t._l(e.tags,(function(e){return n("el-tag",{key:e.id,staticClass:"mr-1 mb-1",attrs:{size:"small"}},[t._v(t._s(e.name))])})),1)],1)}))],2)])}),[],!1,null,null,null).exports},methods:{updateMessage:function(t){this.searchQuery=t}}},z=Object(f.a)(F,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"notes"}},[e("navigation"),this._v(" "),e("div",{staticClass:"row justify-content-center"},[e("algolia-search",{on:{inputData:this.updateMessage}})],1),this._v(" "),e("div",{staticClass:"row"},[e("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[e("notes-list",{attrs:{"search-query":this.searchQuery},on:{inputData:this.updateMessage}})],1)])],1)}),[],!1,null,null,null).exports,H={methods:{textareaCharInserter:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"  ";t.preventDefault();var n=this.note.body,s=t.target.selectionStart,a=n.slice(0,s),r=n.slice(s);this.note.body="".concat(a).concat(e).concat(r),t.target.value=this.note.body,t.target.selectionEnd=t.target.selectionStart=s+1}}};function A(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}function M(t){return function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){A(r,s,a,o,i,"next",t)}function i(t){A(r,s,a,o,i,"throw",t)}o(void 0)}))}}var V={name:"NotesCreate",components:{Navigation:O},mixins:[H],data:function(){return{newTagInputVisible:!1,newTag:"",errors:[],note:{body:""},tags:[],selectedTags:[],inputEl:null}},methods:{insertCodeBlockBackTicks:function(){if(this.inputEl||(this.inputEl=document.querySelector("#note-create textarea")),!document.execCommand("insertText",!1,"```")&&"function"==typeof this.inputEl.setRangeText){var t=this.inputEl.selectionStart;this.inputEl.setRangeText("```"),this.inputEl.selectionStart=this.inputEl.selectionEnd=t+"```".length;var e=document.createEvent("UIEvent");e.initEvent("input",!0,!1),this.inputEl.dispatchEvent(e)}},addNewTag:function(){var t=this;return M(i.a.mark((function e(){var n,s,a,r;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(!t.newTag){e.next=15;break}return t.errors=[],e.prev=2,e.next=5,g.a.post("/api/notetagcreate",{tag:t.newTag});case 5:n=e.sent,s=l()(n,"data.data.success"),a=l()(n,"data.data.id"),s&&a&&(t.tags.push({id:a,name:t.newTag}),t.newTagInputVisible=!1,t.newTag=""),e.next=15;break;case 11:e.prev=11,e.t0=e.catch(2),r=l()(e.t0,"response.data.errors.tag[0]"),t.errors.push({field:"tagCreate",text:r});case 15:case"end":return e.stop()}}),e,null,[[2,11]])})))()},showInput:function(){var t=this;this.newTagInputVisible=!0,this.$nextTick((function(e){t.$refs.saveTagInput.$refs.input.focus()}))},fetchTags:function(){var t=this;return M(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,g.a.get("/api/notes/tags");case 2:n=e.sent,t.tags=l()(n,"data.data",[]);case 4:case"end":return e.stop()}}),e)})))()},submitForm:function(t){var e=this;return M(i.a.mark((function n(){var s,a,r;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t.preventDefault(),!e.newTagInputVisible){n.next=3;break}return n.abrupt("return",!1);case 3:if(e.errors=[],e.note.body){n.next=8;break}return e.errors.push({field:"body",text:"Note text is empty."}),e.note.body=null,n.abrupt("return");case 8:return n.prev=8,n.next=11,g.a.post("/api/notecreate",{body:e.note.body,tags:e.selectedTags});case 11:s=n.sent,a=l()(s,"data.data.id"),r=l()(s,"data.data.success"),a&&r?(e.note.body="",e.selectedTags=[],e.$message({type:"success",message:e.$I18n.trans("notes.confirmation_success_note_create"),center:!0}),setTimeout((function(){return e.$router.push(Object.assign({},Dt,{params:{id:a}}))}),400)):e.errors.push({field:"na",text:"Failed to create the note."}),n.next=20;break;case 17:n.prev=17,n.t0=n.catch(8),e.errors.push({field:"submit",text:n.t0.message});case 20:case"end":return n.stop()}}),n,null,[[8,17]])})))()}},created:function(){this.fetchTags()}},Q=(n("EPOQ"),Object(f.a)(V,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"notes-create"}},[n("navigation"),t._v(" "),n("div",{staticClass:"row justify-content-center"},[n("div",{staticClass:"col-sm no-gutter-xs"},[n("div",{staticClass:"card shadow"},[t.errors.length?n("div",{staticClass:"p-1"},t._l(t.errors,(function(t){return n("el-alert",{key:t.field,attrs:{title:t.text,type:"error"}})})),1):t._e(),t._v(" "),n("form",{attrs:{method:"post"},on:{submit:t.submitForm}},[n("div",{staticClass:"textarea-container"},[n("textarea",{directives:[{name:"model",rawName:"v-model",value:t.note.body,expression:"note.body"}],staticClass:"form-control form-text m-0",attrs:{placeholder:t.$I18n.trans("notes.placeholder_note_title"),name:"body",cols:"100",rows:"15"},domProps:{value:t.note.body},on:{keydown:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"tab",9,e.key,"Tab")?null:(e.preventDefault(),t.textareaCharInserter(e))},input:function(e){e.target.composing||t.$set(t.note,"body",e.target.value)}}})]),t._v(" "),n("div",{staticClass:"d-flex justify-content-between"},[n("small",{staticClass:"form-text text-muted ml-3"},[t._m(0)]),t._v(" "),n("el-button",{attrs:{circle:"",size:"small",type:"info"},on:{click:t.insertCodeBlockBackTicks}},[t._v("```")])],1),t._v(" "),n("div",{staticClass:"mt-3 ml-3"},[n("p",[t._v(t._s(t.$I18n.trans("notes.tags"))+":")]),t._v(" "),n("el-checkbox-group",{staticClass:"d-inline-block",attrs:{size:"small"},model:{value:t.selectedTags,callback:function(e){t.selectedTags=e},expression:"selectedTags"}},t._l(t.tags,(function(e){return n("el-checkbox-button",{key:e.id,staticClass:"mr-1",attrs:{label:e.id}},[t._v(t._s(e.name))])})),1),t._v(" "),n("div",{staticClass:"d-inline-block new-tag-group"},[t.newTagInputVisible?n("el-input",{ref:"saveTagInput",staticClass:"input-new-tag",attrs:{size:"small"},on:{blur:t.addNewTag},nativeOn:{keyup:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"enter",13,e.key,"Enter")?null:t.addNewTag(e)}},model:{value:t.newTag,callback:function(e){t.newTag=e},expression:"newTag"}}):n("el-button",{staticClass:"button-new-tag",attrs:{type:"info",plain:"",size:"small"},on:{click:t.showInput}},[t._v(t._s(t.$I18n.trans("notes.new_tag")))])],1)],1),t._v(" "),n("div",{staticClass:"px-3 pb-3 pt-3"},[n("button",{staticClass:"btn btn-primary",attrs:{type:"submit"}},[t._v(t._s(t.$I18n.trans("common.add")))])])])])])])],1)}),[function(){var t=this.$createElement;return(this._self._c||t)("a",{staticClass:"text-muted",attrs:{href:"https://commonmark.org/help/",tabindex:"-1",target:"_CommonMark"}},[this._v(this._s(this.$I18n.trans("notes.commonmark")))])}],!1,null,"c5e2bf76",null).exports);n("WxZA");function L(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}function q(t){return function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){L(r,s,a,o,i,"next",t)}function i(t){L(r,s,a,o,i,"throw",t)}o(void 0)}))}}var Y={name:"NotesShow",components:{Navigation:O},props:["id"],data:function(){return{loading:!0,note:{},notesUpdateRoute:Ut}},created:function(){this.fetchNote(this.id)},methods:{fetchNote:function(t){var e=this;return q(i.a.mark((function n(){var s;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t){n.next=2;break}return n.abrupt("return",!1);case 2:return n.prev=2,n.next=5,g.a.get("/api/notes/".concat(t,"?html=1"));case 5:s=n.sent,n.next=14;break;case 8:if(n.prev=8,n.t0=n.catch(2),404!==n.t0.response.status){n.next=14;break}return n.next=13,e.$router.push(Object.assign({},Qt,{params:{message:"Note could not be found.",linksTo:[{route:Ot,linkText:"Notes"}]}}));case 13:return n.abrupt("return");case 14:e.note=l()(s,"data.data"),setTimeout((function(){return e.loading=!1}),400);case 16:case"end":return n.stop()}}),n,null,[[2,8]])})))()},confirmDelete:function(t){var e=this;return q(i.a.mark((function n(){var s,a,r,o;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:s=e.$I18n.trans("notes.confirmation_ask_delete",{id:t}),a=e.$I18n.trans("notes.confirmation_success_deleted",{id:t}),r=e.$I18n.trans("notes.delete"),o=e.$I18n.trans("notes.cancel"),e.$confirm(s,"Warning",{confirmButtonText:r,cancelButtonText:o,type:"warning",center:!0}).then((function(){e.deleteNote(t).then((function(t){t&&(e.$message({type:"success",message:a,center:!0}),setTimeout((function(){return e.$router.push(Ot)}),400))})).catch((function(n){e.$message({type:"warning",message:e.$I18n.trans("notes.failed_delete",{id:t}),center:!0})}))})).catch((function(){e.$message({type:"info",message:e.$I18n.trans("notes.delete_canceled",{id:t}),center:!0})}));case 5:case"end":return n.stop()}}),n)})))()},deleteNote:function(t){return q(i.a.mark((function e(){var n,s;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,g.a.delete("/api/notedestroy/".concat(t),{_method:"delete"});case 2:if(n=e.sent,!(s=l()(n,"data.data.success",!1))){e.next=6;break}return e.abrupt("return",s);case 6:throw"Delete failed";case 7:case"end":return e.stop()}}),e)})))()}}},W=(n("WTCb"),Object(f.a)(Y,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"note"},[n("navigation"),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"row justify-content-center"},[n("div",{staticClass:"col-12 no-gutter-xs"},[t.note.id?t._e():n("div",{staticClass:"mt-3 "},[n("p",{staticClass:"loader-text pb-2 pt-3"},[t._v(" ")]),t._v(" "),n("p",{staticClass:"loader-text"},[t._v(" ")]),t._v(" "),n("p",{staticClass:"loader-text"},[t._v(" ")]),t._v(" "),n("p",{staticClass:"loader-text loader-text--1x3 d-inline-block"},[t._v("   ")])]),t._v(" "),n("div",{staticClass:"mt-3",domProps:{innerHTML:t._s(t.note.body)}}),t._v(" "),n("div",t._l(t.note.tags,(function(e){return n("el-tag",{key:e.id,staticClass:"mr-1 mb-1",attrs:{size:"big"}},[t._v(t._s(e.name))])})),1)])]),t._v(" "),n("div",{staticClass:"row"},[n("div",{staticClass:"col-12"},[n("div",{staticClass:"mt-3"},[t.note.id?n("router-link",{staticClass:"btn btn-primary",attrs:{to:Object.assign({},t.notesUpdateRoute,{params:{id:t.note.id}})}},[t._v(t._s(t.$I18n.trans("notes.edit")))]):t._e(),t._v(" "),n("button",{staticClass:"btn btn-danger ml-3",on:{click:function(e){return t.confirmDelete(t.note.id)}}},[t._v(t._s(t.$I18n.trans("notes.delete")))])],1)])])],1)}),[],!1,null,"1e7e8ff5",null).exports);function J(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}function X(t){return function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){J(r,s,a,o,i,"next",t)}function i(t){J(r,s,a,o,i,"throw",t)}o(void 0)}))}}var Z={name:"NotesUpdate",components:{Navigation:O},mixins:[H],props:["id"],data:function(){return{loading:!0,newTagInputVisible:!1,newTag:"",errors:[],note:{},tags:[{id:1,name:""}],selectedTags:[],inputEl:null}},methods:{insertCodeBlockBackTicks:function(){if(this.inputEl||(this.inputEl=document.querySelector("#note-update textarea")),!document.execCommand("insertText",!1,"```")&&"function"==typeof this.inputEl.setRangeText){var t=this.inputEl.selectionStart;this.inputEl.setRangeText("```"),this.inputEl.selectionStart=this.inputEl.selectionEnd=t+"```".length;var e=document.createEvent("UIEvent");e.initEvent("input",!0,!1),this.inputEl.dispatchEvent(e)}},addNewTag:function(){var t=this;return X(i.a.mark((function e(){var n,s,a,r;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(!t.newTag){e.next=15;break}return t.errors=[],e.prev=2,e.next=5,g.a.post("/api/notetagcreate",{tag:t.newTag});case 5:n=e.sent,s=l()(n,"data.data.success"),a=l()(n,"data.data.id"),s&&a&&(t.tags.push({id:a,name:t.newTag}),t.newTagInputVisible=!1,t.newTag=""),e.next=15;break;case 11:e.prev=11,e.t0=e.catch(2),r=l()(e.t0,"response.data.errors.tag[0]"),t.errors.push({field:"tagCreate",text:r});case 15:case"end":return e.stop()}}),e,null,[[2,11]])})))()},showInput:function(){var t=this;this.newTagInputVisible=!0,this.$nextTick((function(e){t.$refs.saveTagInput.$refs.input.focus()}))},fetchNote:function(t){var e=this;return X(i.a.mark((function n(){var s;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t){n.next=2;break}return n.abrupt("return",!1);case 2:return n.next=4,g.a.get("/api/notes/".concat(t,"?html=0"));case 4:s=n.sent,e.note=l()(s,"data.data",[]),e.selectedTags=e.note.tags.map((function(t){return t.id})),setTimeout((function(){return e.loading=!1}),400);case 8:case"end":return n.stop()}}),n)})))()},fetchTags:function(){var t=this;return X(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,g.a.get("/api/notes/tags");case 2:n=e.sent,t.tags=l()(n,"data.data",[]);case 4:case"end":return e.stop()}}),e)})))()},submitForm:function(t){var e=this;return X(i.a.mark((function n(){var s;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t.preventDefault(),!e.newTagInputVisible){n.next=3;break}return n.abrupt("return",!1);case 3:if(e.errors=[],e.note.body){n.next=7;break}return e.errors.push({field:"body",text:"Note text is empty."}),n.abrupt("return");case 7:return n.prev=7,n.next=10,g.a.put("/api/noteupdate/".concat(e.note.id),{body:e.note.body,tags:e.selectedTags,_method:"put"});case 10:s=n.sent,l()(s,"data.data.success")?(e.selectedTags=[],e.$message({type:"success",message:e.$I18n.trans("notes.confirmation_success_note_updated"),center:!0}),setTimeout((function(){return e.$router.push(Object.assign({},Dt,{params:{id:e.note.id}}))}),400)):e.errors.push({field:"na",text:"Failed to update note."}),n.next=17;break;case 14:n.prev=14,n.t0=n.catch(7),e.errors.push({field:"submit",text:n.t0.message});case 17:case"end":return n.stop()}}),n,null,[[7,14]])})))()}},created:function(){var t=this;this.fetchTags().then((function(){t.fetchNote(t.id)}))}},K=(n("O//m"),Object(f.a)(Z,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"notes-update"}},[n("navigation"),t._v(" "),n("div",{staticClass:"row justify-content-center"},[n("div",{staticClass:"col-sm no-gutter-xs"},[n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"card shadow"},[t.errors.length?n("div",{staticClass:"p-1"},t._l(t.errors,(function(t){return n("el-alert",{key:t.field,attrs:{title:t.text,type:"error"}})})),1):t._e(),t._v(" "),n("form",{attrs:{method:"post"},on:{submit:t.submitForm}},[n("div",{staticClass:"textarea-container"},[n("textarea",{directives:[{name:"model",rawName:"v-model",value:t.note.body,expression:"note.body"}],staticClass:"form-control form-text m-0",attrs:{placeholder:t.$I18n.trans("notes.placeholder_note_title"),name:"body",cols:"100",rows:"15"},domProps:{value:t.note.body},on:{keydown:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"tab",9,e.key,"Tab")?null:(e.preventDefault(),t.textareaCharInserter(e))},input:function(e){e.target.composing||t.$set(t.note,"body",e.target.value)}}})]),t._v(" "),n("div",{staticClass:"d-flex justify-content-between"},[n("small",{staticClass:"form-text text-muted ml-3"},[t._m(0)]),t._v(" "),n("el-button",{attrs:{circle:"",size:"small",type:"info"},on:{click:t.insertCodeBlockBackTicks}},[t._v("```")])],1),t._v(" "),n("div",{staticClass:"mt-3 ml-3"},[n("p",[t._v("Tags:")]),t._v(" "),n("el-checkbox-group",{staticClass:"d-inline-block",attrs:{size:"small"},model:{value:t.selectedTags,callback:function(e){t.selectedTags=e},expression:"selectedTags"}},t._l(t.tags,(function(e){return n("el-checkbox-button",{key:e.id,staticClass:"mr-1",attrs:{label:e.id}},[t._v(t._s(e.name))])})),1),t._v(" "),n("div",{staticClass:"d-inline-block new-tag-group"},[t.newTagInputVisible?n("el-input",{ref:"saveTagInput",staticClass:"input-new-tag",attrs:{size:"small"},on:{blur:t.addNewTag},nativeOn:{keyup:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"enter",13,e.key,"Enter")?null:t.addNewTag(e)}},model:{value:t.newTag,callback:function(e){t.newTag=e},expression:"newTag"}}):n("el-button",{staticClass:"button-new-tag",attrs:{type:"info",plain:"",size:"small"},on:{click:t.showInput}},[t._v(t._s(t.$I18n.trans("notes.new_tag")))])],1)],1),t._v(" "),n("div",{staticClass:"px-3 pb-3 pt-3"},[n("button",{staticClass:"btn btn-primary",attrs:{type:"submit"}},[t._v(t._s(t.$I18n.trans("common.update")))])])])])])])],1)}),[function(){var t=this.$createElement;return(this._self._c||t)("a",{staticClass:"text-muted",attrs:{href:"https://commonmark.org/help/",tabindex:"-1",target:"_CommonMark"}},[this._v(this._s(this.$I18n.trans("notes.commonmark")))])}],!1,null,"4b83dcc8",null).exports),G={name:"NotFound",props:{message:{type:String,required:!1,default:"We could not find what you where looking for!"},linksTo:{type:Array,required:!1}},data:function(){return{dashboardRoute:Pt}}},tt=Object(f.a)(G,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"pt-4",attrs:{id:"NotFound"}},[n("h1",{staticClass:"text-center text-capitalize"},[t._v("404 - page not found")]),t._v(" "),n("h5",{staticClass:"text-center text-black-50 mt-4"},[t._v(t._s(t.message))]),t._v(" "),n("div",{staticClass:"px-5 pt-3 text-center"},[n("router-link",{staticClass:"btn btn-lg btn-link",attrs:{to:t.dashboardRoute}},[t._v("Home")]),t._v(" "),t._l(t.linksTo,(function(e){return n("router-link",{key:e.route.name,staticClass:"btn btn-lg btn-link",attrs:{to:e.route}},[t._v(t._s(t.route.linkText))])}))],2)])}),[],!1,null,"98208f0c",null).exports,et={name:"Navigation",props:{numberOfBookmarks:{type:null|Number,required:!0}},data:function(){return{hackerNewsTopPostsRoute:Rt,hackerNewsBestPostsRoute:Ft,hackerNewsBookmarkedPostsRoute:zt}}},nt=(n("JVNZ"),Object(f.a)(et,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100 btn-secondary",attrs:{to:t.hackerNewsTopPostsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.top")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 btn-secondary",attrs:{to:t.hackerNewsBestPostsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.best")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 btn-secondary",attrs:{to:t.hackerNewsBookmarkedPostsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.bookmarks"))+" "),n("span",{staticClass:"badge badge-light"},[t._v(t._s(t.numberOfBookmarks))])])],1)])])])}),[],!1,null,"be214f70",null).exports),st={name:"HackerNews",components:{Navigation:nt},data:function(){return{numberOfBookmarks:null}},methods:{updateNumberOfBookmarks:function(t){this.numberOfBookmarks=t}}},at=Object(f.a)(st,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"hacker-news"}},[e("navigation",{attrs:{"number-of-bookmarks":this.numberOfBookmarks}}),this._v(" "),e("router-view",{on:{nBookmarksChangedEvent:this.updateNumberOfBookmarks}})],1)}),[],!1,null,null,null).exports,rt={name:"HnComment",props:{comment:{type:Object,required:!0},handleCollapseToggle:{type:Function,required:!0}},data:function(){return{isShow:!0}}},ot=(n("P9EP"),Object(f.a)(rt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"hn-comment card mt-3"},[n("div",{staticClass:"card-body px-2 px-md-3 shadow"},[n("h6",{staticClass:"card-subtitle bg-light text-muted d-flex justify-content-between align-items-center",class:{"opacity-5 text-line-through":t.comment.deleted}},[t.comment.id?n("span",{staticClass:"align-middle"},[t._v(t._s(t.$I18n.trans("hackernews.by",{by:t.comment.by||""})))]):t._e(),t._v(" "),n("small",{staticClass:"text-muted",attrs:{title:t.comment.time}},[t._v(t._s(t._f("diffForHumans")(t.comment.time)))]),t._v(" "),t.comment.id?n("a",{staticClass:"btn btn-sm pointer btn-collapse",attrs:{role:"button","data-toggle":"collapse"},on:{click:function(e){return t.handleCollapseToggle(t.comment)}}},[n("b",[t._v("["),t.comment.collapsed?n("span",[t._v(t._s(t.comment.kids.length))]):n("span",[t._v("-")]),t._v("]\n        ")])]):t._e()]),t._v(" "),n("transition",{attrs:{name:"slide"}},[t.comment.collapsed?t._e():n("div",{staticClass:"mt-2"},[n("p",{domProps:{innerHTML:t._s(t.comment.text)}}),t._v(" "),t._l(t.comment.comments,(function(e){return t.comment.comments.length?n("hn-comment",{key:e.id,attrs:{comment:e,"handle-collapse-toggle":t.handleCollapseToggle}}):t._e()}))],2)])],1)])}),[],!1,null,"1a92697e",null).exports);function it(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var ct={data:function(){return{nBookmarks:null,notification:null}},methods:{bookmarkPost:function(t){var e,n=this;return(e=i.a.mark((function e(){var s,a,r,o;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return t.status.saving=!0,s={postId:t.id},(a=!t.status.bookmarked)||(s._method="delete"),e.next=6,g.a.post("/api/bookmarks",s);case 6:r=e.sent,(o=l()(r,"data.data.success",!1))&&(t.status.bookmarked=!t.status.bookmarked,n.nBookmarks+=a?1:-1),n.notifyUserOfBookmarkStatusChange(o,a,t);case 10:case"end":return e.stop()}}),e)})),function(){var t=this,n=arguments;return new Promise((function(s,a){var r=e.apply(t,n);function o(t){it(r,s,a,o,i,"next",t)}function i(t){it(r,s,a,o,i,"throw",t)}o(void 0)}))})()},notifyUserOfBookmarkStatusChange:function(t,e,n){var s={message:this.$I18n.trans("hackernews.add_failure"),type:"error"};t&&(s=e?{message:this.$I18n.trans("hackernews.added_to_bookmarks",{title:n.title}),type:"success"}:{message:this.$I18n.trans("hackernews.remove_from_bookmarks",{title:n.title}),type:"warning"}),this.notification&&!this.notification.closed&&this.notification.close(),this.notification=this.$message(s),setTimeout((function(){return n.status.saving=!1}),400)}}},lt=n("Wcq6"),ut=n.n(lt),dt=(n("Zs65"),ut.a.initializeApp({databaseURL:"hacker-news.firebaseio.com"}).database().ref("/v0"));function mt(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var pt={name:"HackerNewsPost",mixins:[ct],components:{Navigation:nt,HnComment:ot},props:{id:{type:Number|String,required:!0}},data:function(){return{loading:!0,loadingComments:!0,collapsedCommentsKey:"collapsedCommentsHackerNewsPost",collapsedComments:[],post:{id:null,by:null,title:null,text:null,score:null,time:null,url:null,type:null,nComments:null,status:{bookmarked:null,saving:null},kids:[],comments:[{id:null,by:null,kids:[],parent:null,text:null,time:null,comments:[]}]},numberOfBookmarks:null}},methods:{fetchPost:function(t){var e=this;t&&(this.fetchCollapsedComments(),dt.child("item/".concat(t)).once("value",(function(t){var n=t.val();e.post.id=l()(n,"id"),e.post.by=l()(n,"by"),e.post.title=l()(n,"title"),e.post.text=l()(n,"text"),e.post.score=l()(n,"score"),e.post.time=d.a.unix(l()(n,"time",0)).format("YYYY-MM-DD HH:mm:ss"),e.post.url=l()(n,"url"),e.post.type=l()(n,"type"),e.post.nComments=l()(n,"descendants"),e.post.bookmarked=null,e.post.kids=l()(n,"kids",[]),e.post.comments=[],e.setBookmarkStatus(e.post),e.fetchComments(e.post),e.loading=!1})))},fetchComments:function(t){var e=this;t.kids.length&&t.kids.map((function(n){e.fetchComment(n).then((function(n){var s=l()(n,"kids",[]),a=l()(n,"id"),r={id:a,deleted:l()(n,"deleted"),by:l()(n,"by"),kids:s,text:l()(n,"text"),time:d.a.unix(l()(n,"time",0)).format("YYYY-MM-DD HH:mm:ss"),comments:[],collapsed:e.collapsedComments.includes(a)};e.fetchComments(r),t.comments.push(r),e.loadingComments=!1}))}))},fetchComment:function(t){return new Promise((function(e,n){dt.child("item/".concat(t)).once("value",(function(t){var n=t.val();e(n)}))}))},setBookmarkStatus:function(t){var e,n=this;return(e=i.a.mark((function e(){var s,a;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,g.a.get("/api/bookmarks");case 2:s=e.sent,a=l()(s,"data.data",[]),n.numberOfBookmarks=a.length,t.status.bookmarked=a.includes(t.id);case 6:case"end":return e.stop()}}),e)})),function(){var t=this,n=arguments;return new Promise((function(s,a){var r=e.apply(t,n);function o(t){mt(r,s,a,o,i,"next",t)}function i(t){mt(r,s,a,o,i,"throw",t)}o(void 0)}))})()},fetchCollapsedComments:function(){var t,e=localStorage.getItem(this.collapsedCommentsKey);try{t=JSON.parse(e)}catch(t){console.error(t)}t=Array.isArray(t)?t:[],this.collapsedComments=t},setCollapsedComments:function(){var t=JSON.stringify(this.collapsedComments);localStorage.setItem(this.collapsedCommentsKey,t)},handleCollapseToggle:function(t){t.collapsed=!t.collapsed,t.collapsed?this.collapsedComments.push(t.id):this.collapsedComments=this.collapsedComments.filter((function(e){return e!==t.id})),this.setCollapsedComments()}},created:function(){this.fetchPost(this.id)}},ft=(n("zTjN"),Object(f.a)(pt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"hacker-news-post"}},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}]},[t.loading?n("p",{staticClass:"lead loader-text d-block loader-text--2x3"},[t._v(" ")]):t._e(),t._v(" "),t.post.title?n("p",{staticClass:"lead"},[t._v("\n          "+t._s(t.post.title)+"\n          "),t.post.url?n("a",{staticClass:"text-body",attrs:{href:t.post.url,target:"_blank"}},[n("small",{staticClass:"text-muted"},[t._v("("+t._s(t._f("domainFromUrl")(t.post.url))+") [↗]")])]):t._e()]):t._e(),t._v(" "),t.post.text?n("p",{domProps:{innerHTML:t._s(t.post.text)}}):t._e(),t._v(" "),t.loading?n("div",{staticClass:"loader-text d-block loader-text--1x3"},[t._v(" ")]):t._e(),t._v(" "),t.post.title?n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.post.status.saving,expression:"post.status.saving"}]},[n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.points",{points:t.post.score})))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.comments",{comments:t.post.nComments})))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("small",{staticClass:"text-muted",attrs:{title:t.post.time}},[t._v(t._s(t._f("diffForHumans")(t.post.time)))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.by",{by:t.post.by})))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("span",{staticClass:"bookmark-story pointer text-primary",on:{click:function(e){return t.bookmarkPost(t.post)}}},[t._v(t._s(t.post.status.bookmarked?"⚫":"⚪️")+"️")])]):t._e()])])]),t._v(" "),t.post.comments.length?n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 no-gutter-xs"},[t.loadingComments?t._e():n("div",t._l(t.post.comments,(function(e){return n("hn-comment",{key:e.id,attrs:{comment:e,"handle-collapse-toggle":t.handleCollapseToggle}})})),1)])]):t._e()])}),[],!1,null,"cdd94d26",null).exports),vt={name:"Navigation",data:function(){return{usersRoute:Mt,usersCreateRoute:Vt}},methods:{activeSubmenu:function(t){return this.$route.name===t?"btn-primary":"btn-secondary"}}},ht=Object(f.a)(vt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100",class:t.activeSubmenu("Users"),attrs:{to:t.usersRoute}},[t._v(t._s(t.$I18n.trans("users.users")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100",class:t.activeSubmenu("UsersCreate"),attrs:{to:t.usersCreateRoute}},[t._v(t._s(t.$I18n.trans("users.new-user")))])],1)])])])}),[],!1,null,null,null).exports;function gt(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}function bt(t){return function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){gt(r,s,a,o,i,"next",t)}function i(t){gt(r,s,a,o,i,"throw",t)}o(void 0)}))}}var _t={name:"Users",components:{Navigation:ht},data:function(){return{loading:!0,users:[],userToEdit:{},currentUserId:null,dialogVisible:!1,dialogFormRules:{name:[{required:!0,message:this.$I18n.trans("users.name_cannot_be_empty"),trigger:"blur"},{min:2,max:255,message:this.$I18n.trans("users.min_2_max_255"),trigger:"blur"}]}}},methods:{fetchUsers:function(){var t=this;return bt(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,g.a.get("/api/users");case 2:n=e.sent,t.users=l()(n,"data.data.users",[]),t.currentUserId=l()(n,"data.data.currentUserId",[]),t.loading=!1;case 6:case"end":return e.stop()}}),e)})))()},highlightCurrentUser:function(t){var e=t.row;t.rowIndex;return e.id===this.currentUserId?"table-success":null},formatDate:function(t){return this.$options.filters.diffForHumans(t)},formatCreatedAt:function(t,e){return this.formatDate(t.createdAt)},formatUpdatedAt:function(t,e){return this.formatDate(t.updatedAt)},closeDialog:function(t){this.dialogVisible=!1,this.$refs[t].resetFields()},editDialogShow:function(t){this.userToEdit={id:t.id,name:t.name,email:t.email},this.dialogVisible=!0},editDialogSubmit:function(t){var e=this;this.$refs[t].validate((function(t){return!!t&&(e.updateUser(),!0)}))},updateUser:function(){var t=this;return bt(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,e.next=3,g.a.put("/api/users/update",{id:t.userToEdit.id,name:t.userToEdit.name});case 3:n=e.sent,e.next=11;break;case 6:return e.prev=6,e.t0=e.catch(0),console.error(e.t0),t.$message.error(t.$I18n.trans("users.failed_to_update")),e.abrupt("return");case 11:t.notifyActionSuccess(n);case 12:case"end":return e.stop()}}),e,null,[[0,6]])})))()},notifyActionSuccess:function(t){var e=this,n=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"users.user_updated",s=arguments.length>2&&void 0!==arguments[2]?arguments[2]:"users.failed_to_update",a=l()(t,"status"),r=l()(t,"data.data.success",!1);if(200===a&&r){this.$message.success(this.$I18n.trans(n));var o=this.users.find((function(t){return t.id===e.userToEdit.id}));o.name=this.userToEdit.name.trim(),this.dialogVisible=!1}else this.$message.error(this.$I18n.trans(s))},deleteUser:function(){var t=this;return bt(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,t.askUserToConfirmDelete();case 2:if(!e.sent){e.next=18;break}return e.prev=4,e.next=7,g.a.delete("/api/users/destroy",{data:{id:t.userToEdit.id}});case 7:n=e.sent,console.log(n),e.next=16;break;case 11:return e.prev=11,e.t0=e.catch(4),console.error(e.t0),t.$message.error(t.$I18n.trans("users.failed_to_delete")),e.abrupt("return");case 16:t.notifyActionSuccess(n,"users.delete_success","users.failed_to_delete"),t.fetchUsers();case 18:case"end":return e.stop()}}),e,null,[[4,11]])})))()},askUserToConfirmDelete:function(){var t=this;return bt(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,e.next=3,t.$confirm(t.$I18n.trans("users.confirm_delete_request",{name:t.userToEdit.name}),t.$I18n.trans("common.please_confirm"),{confirmButtonText:t.$I18n.trans("users.delete"),cancelButtonText:t.$I18n.trans("common.cancel"),type:"warning"});case 3:e.t0=e.sent,n="confirm"===e.t0,e.next=10;break;case 7:e.prev=7,e.t1=e.catch(0),n=!1;case 10:return e.abrupt("return",n);case 11:case"end":return e.stop()}}),e,null,[[0,7]])})))()}},created:function(){this.fetchUsers()}},xt=(n("Od/V"),Object(f.a)(_t,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"users"}},[n("navigation"),t._v(" "),n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[n("el-table",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticStyle:{width:"100%"},attrs:{data:t.users,"row-class-name":t.highlightCurrentUser,"empty-text":t.$I18n.trans("users.failed_to_obtain_user_list")}},[n("el-table-column",{attrs:{prop:"id",label:"#"}}),t._v(" "),n("el-table-column",{attrs:{prop:"name",label:t.$I18n.trans("users.name")}}),t._v(" "),n("el-table-column",{attrs:{prop:"email",label:t.$I18n.trans("users.email")}}),t._v(" "),n("el-table-column",{attrs:{prop:"createdAt",label:t.$I18n.trans("users.created_at"),formatter:t.formatCreatedAt}}),t._v(" "),n("el-table-column",{attrs:{prop:"updatedAt",label:t.$I18n.trans("users.updated_at"),formatter:t.formatUpdatedAt}}),t._v(" "),n("el-table-column",{attrs:{fixed:"right",align:"right",width:"60",prop:"updatedAt"},scopedSlots:t._u([{key:"default",fn:function(e){return[n("el-button",{attrs:{type:"text",size:"small"},on:{click:function(n){return t.editDialogShow(e.row)}}},[t._v(t._s(t.$I18n.trans("users.edit")))])]}}])})],1)],1)]),t._v(" "),n("el-dialog",{attrs:{title:t.$I18n.trans("users.user_edit",{id:t.userToEdit.id||""}),visible:t.dialogVisible},on:{"update:visible":function(e){t.dialogVisible=e}}},[n("el-form",{ref:"formRef",attrs:{model:t.userToEdit,rules:t.dialogFormRules},on:{submit:t.editDialogSubmit},nativeOn:{submit:function(t){t.preventDefault()}}},[n("el-form-item",{attrs:{label:t.$I18n.trans("users.name"),prop:"name"}},[n("el-input",{attrs:{placeholder:t.$I18n.trans("users.name_cannot_be_empty")},model:{value:t.userToEdit.name,callback:function(e){t.$set(t.userToEdit,"name",e)},expression:"userToEdit.name"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.email")}},[n("el-input",{attrs:{type:"email",value:t.userToEdit.email,disabled:""}})],1)],1),t._v(" "),n("span",{staticClass:"dialog-footer",attrs:{slot:"footer"},slot:"footer"},[n("el-button",{on:{click:function(e){return t.closeDialog("formRef")}}},[t._v(t._s(t.$I18n.trans("common.cancel")))]),t._v(" "),n("el-button",{attrs:{type:"danger",disabled:t.userToEdit.id===t.currentUserId},on:{click:function(e){return t.deleteUser()}}},[t._v(t._s(t.$I18n.trans("users.delete")))]),t._v(" "),n("el-button",{attrs:{type:"primary"},on:{click:function(e){return t.editDialogSubmit("formRef")}}},[t._v(t._s(t.$I18n.trans("common.confirm")))])],1)],1)],1)}),[],!1,null,"051ebd10",null).exports);function kt(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var yt={name:"UsersCreate",components:{Navigation:ht},data:function(){return{rules:{name:[{required:!0,message:this.$I18n.trans("users.name_cannot_be_empty"),trigger:"blur"},{min:2,max:255,message:this.$I18n.trans("users.min_2_max_255"),trigger:"blur"}],email:[{required:!0,type:"email",message:this.$I18n.trans("users.please_set_valid_email"),trigger:"blur"}],password:[{required:!0,message:this.$I18n.trans("users.please_set_valid_password"),trigger:"blur"},{min:12,max:100,message:this.$I18n.trans("users.please_set_valid_password"),trigger:"blur"}]},user:{name:"",email:"",password:""}}},methods:{submitForm:function(t){var e=this;this.$refs[t].validate((function(t){return!!t&&(e.createUser(),!0)}))},createUser:function(){var t,e=this;return(t=i.a.mark((function t(){var n;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.prev=0,t.next=3,g.a.post("/api/users/create",e.user);case 3:n=t.sent,t.next=9;break;case 6:t.prev=6,t.t0=t.catch(0),console.error(t.t0);case 9:e.notifyActionSuccess(n);case 10:case"end":return t.stop()}}),t,null,[[0,6]])})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){kt(r,s,a,o,i,"next",t)}function i(t){kt(r,s,a,o,i,"throw",t)}o(void 0)}))})()},notifyActionSuccess:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"users.user_create",n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:"users.failed_to_create",s=l()(t,"status"),a=l()(t,"data.data.success",!1);200===s&&a?(this.$message.success(this.$I18n.trans(e)),this.$router.push(Mt)):this.$message.error(this.$I18n.trans(n))}}},wt=Object(f.a)(yt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"users-create"}},[n("navigation"),t._v(" "),n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[n("el-form",{ref:"refCreateUserForm",attrs:{model:t.user,rules:t.rules}},[n("el-form-item",{attrs:{label:t.$I18n.trans("users.name"),prop:"name"}},[n("el-input",{model:{value:t.user.name,callback:function(e){t.$set(t.user,"name",e)},expression:"user.name"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.email"),prop:"email"}},[n("el-input",{attrs:{type:"email",autocomplete:"off"},model:{value:t.user.email,callback:function(e){t.$set(t.user,"email",e)},expression:"user.email"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.password"),prop:"password"}},[n("el-input",{attrs:{type:"password",autocomplete:"off"},model:{value:t.user.password,callback:function(e){t.$set(t.user,"password",e)},expression:"user.password"}})],1),t._v(" "),n("el-form-item",[n("el-button",{attrs:{type:"primary"},on:{click:function(e){return t.createUser("refCreateUserForm")}}},[t._v(t._s(t.$I18n.trans("users.create")))])],1)],1)],1)])],1)}),[],!1,null,"06f558c4",null).exports;function Ct(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}function jt(t){return function(t){if(Array.isArray(t))return It(t)}(t)||function(t){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t))return Array.from(t)}(t)||function(t,e){if(!t)return;if("string"==typeof t)return It(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return It(t,e)}(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function It(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,s=new Array(e);n<e;n++)s[n]=t[n];return s}var $t={name:"HNPosts",mixins:[ct],props:{idList:{type:Array,default:[],required:!0}},data:function(){return{loading:!0,posts:[],nBookmarks:null,firstBatchPosition:10,hackerNewsPostRoute:Ht}},methods:{fetchItems:function(t){var e=this,n=this.firstBatchPosition;t.length<n&&(n=Math.floor(t.length/2));var s=t.slice(0,n),a=t.slice(n);Promise.all(s.map((function(t){return e.fetchItem(t)}))).then((function(t){e.posts=t.filter(e.isStoryValid),e.attachBookmarked(),e.loading=!1})),this.$nextTick((function(){Promise.all(a.map((function(t){return e.fetchItem(t)}))).then((function(t){var n;(n=e.posts).push.apply(n,jt(t.filter(e.isStoryValid))),e.attachBookmarked()}))}))},isStoryValid:function(t){return"story"===t.type&&!!t.id},fetchItem:function(t){return new Promise((function(e,n){dt.child("item/".concat(t)).on("value",(function(t){var n=t.val(),s={id:l()(n,"id"),title:l()(n,"title"),score:l()(n,"score"),time:d.a.unix(l()(n,"time",0)).format("YYYY-MM-DD HH:mm:ss"),url:l()(n,"url"),type:l()(n,"type"),nComments:l()(n,"descendants"),kids:l()(n,"kids",[]),status:{bookmarked:!1,saving:!1}};e(s)})),dt.onDisconnect((function(t,e){this.$alert("Firebase connection lost","Lost connection",{confirmButtonText:"OK"}),console.log(t,e)}))}))},attachBookmarked:function(){var t,e=this;return(t=i.a.mark((function t(){var n,s;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,g.a.get("/api/bookmarks");case 2:n=t.sent,s=l()(n,"data.data",[]),e.nBookmarks=s.length,e.posts=e.posts.map((function(t){return t.status.bookmarked=s.includes(t.id),t}));case 6:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){Ct(r,s,a,o,i,"next",t)}function i(t){Ct(r,s,a,o,i,"throw",t)}o(void 0)}))})()}},watch:{idList:function(t,e){var n=this;this.loading=!0,setTimeout((function(){return n.fetchItems(t)}),400)},nBookmarks:function(t,e){this.$emit("nBookmarksChangedEvent",t)}}};n("0P3G");function Tt(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var Et={name:"HackerNewsList",components:{Navigation:nt,HnPosts:Object(f.a)($t,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("ul",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"list-group",attrs:{id:"hn-posts"}},[t._l(Array(t.firstBatchPosition).fill(null),(function(e){return t.loading?n("li",{staticClass:"list-group-item list-group-item-action flex-column align-items-start"},[t._m(0,!0),t._v(" "),t._m(1,!0)]):t._e()})),t._v(" "),t._l(t.posts,(function(e,s){return n("li",{key:e.id,staticClass:"list-group-item list-group-item-action flex-column align-items-start",attrs:{"data-index":s+1}},[n("div",{staticClass:"d-flex w-100 justify-content-between"},[n("span",[n("router-link",{staticClass:"text-body",attrs:{to:Object.assign({},t.hackerNewsPostRoute,{params:{id:e.id}})}},[t._v(t._s(e.title))]),t._v(" "),e.url?n("a",{staticClass:"text-body",attrs:{href:e.url,target:"_blank"}},[n("small",{staticClass:"text-muted"},[t._v("("+t._s(t._f("domainFromUrl")(e.url))+")")])]):t._e()],1),t._v(" "),n("span",{directives:[{name:"loading",rawName:"v-loading",value:e.status.saving,expression:"post.status.saving"}],staticClass:"d-block d-md-none"},[n("span",{staticClass:"bookmark-story pointer text-primary",on:{click:function(n){return t.bookmarkPost(e)}}},[t._v(t._s(e.status.bookmarked?"⚫":"⚪️")+"️")])]),t._v(" "),n("span",{staticClass:"badge d-none d-md-block",attrs:{title:e.time}},[t._v(t._s(t._f("diffForHumans")(e.time)))])]),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:e.status.saving,expression:"post.status.saving"}],staticClass:"d-none d-md-block"},[n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.points",{points:e.score||""})))]),t._v("\n      |\n      "),n("small",[t._v(t._s(t.$I18n.trans("hackernews.comments",{comments:e.nComments||""})))]),t._v("\n      |\n      "),n("span",{staticClass:"bookmark-story pointer text-primary",on:{click:function(n){return t.bookmarkPost(e)}}},[t._v(t._s(e.status.bookmarked?"⚫":"⚪️"))])])])}))],2)}),[function(){var t=this.$createElement,e=this._self._c||t;return e("div",{staticClass:"d-flex w-100 justify-content-between"},[e("span",{staticClass:"loader-text loader-text-top d-block w-100"},[this._v(" ")])])},function(){var t=this.$createElement,e=this._self._c||t;return e("div",{staticClass:"d-none d-md-block"},[e("span",{staticClass:"loader-text loader-text-bottom loader-text--1x3 d-inline-block"},[e("small",[this._v(" ")])])])}],!1,null,"a624b0ec",null).exports},props:{type:{type:String,validator:function(t){return-1!==["top","best","bookmarks"].indexOf(t)},default:"top",required:!0}},data:function(){return{idList:[],limit:100,types:{HackerNewsBookmarks:"bookmarks"},numberOfBookmarks:null}},methods:{fetchStories:function(){this.type===this.types.HackerNewsBookmarks?this.fetchIdsBookmarkedFromBackend():this.fetchIdsFromFirebase(this.type)},fetchIdsFromFirebase:function(t){var e=this;dt.child("".concat(t,"stories")).limitToFirst(this.limit).once("value",(function(t,n){var s=t.val();e.idList=Array.isArray(s)?s:[]}))},fetchIdsBookmarkedFromBackend:function(){var t,e=this;return(t=i.a.mark((function t(){var n;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,g.a.get("/api/bookmarks");case 2:n=t.sent,e.idList=l()(n,"data.data",[]);case 4:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){Tt(r,s,a,o,i,"next",t)}function i(t){Tt(r,s,a,o,i,"throw",t)}o(void 0)}))})()},updateNumberOfBookmarks:function(t){this.numberOfBookmarks=t}},created:function(){this.fetchStories()},watch:{$route:function(t,e){this.fetchStories()}}},St=Object(f.a)(Et,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"hacker-news-list"}},[e("div",{staticClass:"row"},[e("div",{staticClass:"col-12 no-gutter-xs"},[e("hn-posts",this._g({attrs:{"id-list":this.idList}},this.$listeners))],1)])])}),[],!1,null,null,null).exports;a.a.use(r.a);var Nt={path:"/authenticate",name:"Authenticate",component:N},Pt={path:"/dashboard",alias:"/",name:"Dashboard",component:$},Ot={path:"/notes",name:"Notes",component:z},Bt={path:"/notes/new",name:"NotesCreate",component:Q},Dt={path:"/notes/:id",name:"NotesShow",component:W,props:!0},Ut={path:"/notes/edit/:id",name:"NotesUpdate",component:K,props:!0},Rt={path:"",alias:"top",name:"HackerNewsTopPostsRoute",component:St,props:{type:"top"}},Ft={path:"best",name:"HackerNewsBestPostsRoute",component:St,props:{type:"best"}},zt={path:"bookmarks",name:"HackerNewsBookmarkedPostsRoute",component:St,props:{type:"bookmarks"}},Ht={path:"/:id",name:"HackerNewsPost",component:ft,props:!0},At={path:"/hn",component:at,children:[Rt,Ft,zt,Ht]},Mt={path:"/users",name:"Users",component:xt},Vt={path:"/users/new",name:"UsersCreate",component:wt},Qt={path:"/404",name:"NotFound",component:tt,props:!0},Lt=new r.a({routes:[Nt,Pt,Ot,Bt,Dt,Ut,At,Mt,Vt,Qt,{path:"*",redirect:"/404"}]}),qt=n("XJYT"),Yt=n.n(qt);n("D66Q");function Wt(t){return(Wt="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function Jt(t,e){return function(t){if(Array.isArray(t))return t}(t)||function(t,e){if("undefined"==typeof Symbol||!(Symbol.iterator in Object(t)))return;var n=[],s=!0,a=!1,r=void 0;try{for(var o,i=t[Symbol.iterator]();!(s=(o=i.next()).done)&&(n.push(o.value),!e||n.length!==e);s=!0);}catch(t){a=!0,r=t}finally{try{s||null==i.return||i.return()}finally{if(a)throw r}}return n}(t,e)||function(t,e){if(!t)return;if("string"==typeof t)return Xt(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return Xt(t,e)}(t,e)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function Xt(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,s=new Array(e);n<e;n++)s[n]=t[n];return s}function Zt(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function Kt(t,e){for(var n=0;n<e.length;n++){var s=e[n];s.enumerable=s.enumerable||!1,s.configurable=!0,"value"in s&&(s.writable=!0),Object.defineProperty(t,s.key,s)}}a.a.filter("diffForHumans",(function(t){if(t)return d()(t,"YYYY-MM-DD hh:mm:ss").fromNow()})),a.a.filter("domainFromUrl",(function(t){if(t)return new URL(t).hostname.replace("www.","")}));var Gt=function(){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:"translations";Zt(this,t),this.key=e}var e,n,s;return e=t,(n=[{key:"trans",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{};return this._replace(this._extract(t),e)}},{key:"trans_choice",value:function(t){var e,n=this,s=arguments.length>1&&void 0!==arguments[1]?arguments[1]:1,a=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},r=this._extract(t,"|").split("|");return r.some((function(t){return e=n._match(t,s)})),e=(e=e||(s>1?r[1]:r[0])).replace(/\[.*?\]|\{.*?\}/,""),this._replace(e,a)}},{key:"_match",value:function(t,e){var n=t.match(/^[\{\[]([^\[\]\{\}]*)[\}\]](.*)/);if(n){if(n[1].includes(",")){var s=Jt(n[1].split(",",2),2),a=s[0],r=s[1];if("*"===r&&e>=a)return n[2];if("*"===a&&e<=r)return n[2];if(e>=a&&e<=r)return n[2]}return n[1]==e?n[2]:null}}},{key:"_replace",value:function(t,e){if("object"===Wt(t))return t;for(var n in e)t=t.toString().replace(":".concat(n),e[n]).replace(":".concat(n.toUpperCase()),e[n].toString().toUpperCase()).replace(":".concat(n.charAt(0).toUpperCase()).concat(n.slice(1)),e[n].toString().charAt(0).toUpperCase()+e[n].toString().slice(1));return t.toString().trim()}},{key:"_extract",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:null,n=t.toString().split("::"),s=n.pop().toString().split(".");return n.length>0&&(n[0]+="::"),n.concat(s).reduce((function(n,s){return n[s]||e||t}),window[this.key])}}])&&Kt(e.prototype,n),s&&Kt(e,s),t}(),te=n("D/Jt"),ee={name:"MainNavigation",data:function(){return{notesRoute:Ot,dashboardRoute:Pt,hackerNewsRoute:At,usersRoute:Mt,showMenu:!0,menuCollapsed:!0,appRoutes:[]}},methods:{logout:function(t){var e=this;t.preventDefault(),g.a.post("/logout",{}).then((function(){e.$router.push(Nt),e.showMenu=!1}))},toggleMenuVisibility:function(t){this.showMenu=Nt.name!==t.name}},watch:{$route:function(t,e){this.toggleMenuVisibility(t)}},mounted:function(){this.toggleMenuVisibility(this.$route)}},ne=(n("I/cB"),Object(f.a)(ee,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return t.showMenu?n("nav",{staticClass:"navbar navbar-expand-md navbar-light navbar-laravel",attrs:{id:"main-navigation"}},[n("div",{staticClass:"container"},[n("router-link",{staticClass:"navbar-brand",attrs:{exact:"",to:t.dashboardRoute}},[t._v("Zetetic Elench")]),t._v(" "),n("button",{staticClass:"navbar-toggler",attrs:{type:"button"},on:{click:function(e){t.menuCollapsed=!t.menuCollapsed}}},[n("span",{staticClass:"navbar-toggler-icon"})]),t._v(" "),n("div",{staticClass:"navbar-collapse",class:{collapse:t.menuCollapsed}},[n("ul",{staticClass:"navbar-nav mr-auto"},[n("li",{staticClass:"nav-item"},[n("router-link",{staticClass:"nav-link text-right text-sm-left",attrs:{exact:"",to:t.notesRoute}},[t._v(t._s(t.$I18n.trans("notes.notes")))])],1),t._v(" "),n("li",{staticClass:"nav-item"},[n("router-link",{staticClass:"nav-link text-right text-sm-left",attrs:{to:t.hackerNewsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.hackernews")))])],1),t._v(" "),n("li",{staticClass:"nav-item"},[n("router-link",{staticClass:"nav-link text-right text-sm-left",attrs:{exact:"",to:t.usersRoute}},[t._v(t._s(t.$I18n.trans("users.users")))])],1)]),t._v(" "),n("ul",{staticClass:"navbar-nav ml-auto"},[n("li",{staticClass:"nav-item"},[n("a",{staticClass:"nav-link text-right text-sm-left",attrs:{href:"#"},on:{click:function(e){return t.logout(e)}}},[t._v(t._s(t.$I18n.trans("users.logout")))])])])])],1)]):t._e()}),[],!1,null,"01da79fe",null).exports);a.a.component("main-navigation",ne),a.a.use(Yt.a),a.a.use(te.a),a.a.prototype.$I18n=new Gt,g.a.defaults.headers.common["X-Requested-With"]="XMLHttpRequest",g.a.defaults.withCredentials=!0,g.a.interceptors.response.use(null,(function(t){if(t.response)return 401!==t.response.status&&419!==t.response.status||se.$router.push(Nt),Promise.reject(t);console.error("Unknown error: ".concat(t))}));var se=new a.a({router:Lt,el:"#app"})},fysm:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.bookmark-story[data-v-a624b0ec] {\n  position: relative;\n  top: 2px;\n  text-decoration: none;\n}\n.loader-text[data-v-a624b0ec] {\n  background: #808080;\n  border-radius: 2px;\n}\n.loader-text--1x3[data-v-a624b0ec] {\n  width: 33%;\n}\n.loader-text-top[data-v-a624b0ec] {\n  height: 22px;\n  margin-bottom: 6px;\n}\n.loader-text-bottom[data-v-a624b0ec] {\n  height: 18px;\n}\n",""])},hW3y:function(t,e,n){var s=n("a5qU");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},iEU0:function(t,e,n){var s=n("Boyg");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},k1kw:function(t,e,n){var s=n("7Pbq");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},maHV:function(t,e,n){"use strict";var s=n("iEU0");n.n(s).a},pyCd:function(t,e){},shRJ:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.router-link-exact-active[data-v-be214f70] {\n  color: #fff;\n  background-color: #3490dc;\n  border-color: #3490dc;\n}\n.router-link-exact-active[data-v-be214f70]:focus, .router-link-exact-active[data-v-be214f70]:hover {\n  color: #fff;\n  background-color: #227dc7;\n  border-color: #2176bd;\n}\n",""])},veFt:function(t,e,n){var s=n("WobJ");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},zTjN:function(t,e,n){"use strict";var s=n("D1a/");n.n(s).a}},[[0,1,2]]]);
+(window.webpackJsonp=window.webpackJsonp||[]).push([[0],{"/BTM":function(t,e,n){var s=n("fysm");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},0:function(t,e,n){n("bUC5"),t.exports=n("pyCd")},"0P3G":function(t,e,n){"use strict";var s=n("/BTM");n.n(s).a},"2LfK":function(t,e,n){"use strict";var s=n("lLLT");n.n(s).a},4101:function(t,e,n){"use strict";var s=n("IEyY");n.n(s).a},"7HpC":function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\ntextarea[data-v-4b83dcc8] {\n  outline: 0;\n  background: #f4f4f4;\n  box-shadow: 1px 1px 0 #ccc;\n  border-color: transparent;\n  border-radius: 0;\n  font-family: 'Inconsolata', monospace;\n}\ntextarea[data-v-4b83dcc8]:focus {\n  box-shadow: 1px 1px 0 #ccc;\n  outline: 0;\n  background: #f4f4f4;\n  border-color: transparent;\n}\n.new-tag-group[data-v-4b83dcc8] {\n  font-size: 0;\n}\n.new-tag-group .input-new-tag[data-v-4b83dcc8], .new-tag-group .button-new-tag[data-v-4b83dcc8] {\n  vertical-align: middle;\n  width: 120px;\n}\n#note-update[data-v-4b83dcc8] .el-checkbox-button--small .el-checkbox-button__inner {\n  border-radius: 4px;\n  border: 1px solid #DCDFE6;\n}\n#note-update[data-v-4b83dcc8] .el-checkbox-button.is-checked .el-checkbox-button__inner {\n  border-color: #409EFF;\n}\n",""])},"7Pbq":function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,'\n.loader-text[data-v-1e7e8ff5] {\n  background: #808080;\n  border-radius: 2px;\n}\n.loader-text--1x3[data-v-1e7e8ff5] {\n  width: 33%;\n}\n.note[data-v-1e7e8ff5] .footnotes {\n  font-size: 0.8rem;\n}\n.note[data-v-1e7e8ff5] .footnotes p {\n  margin-bottom: 0.5rem;\n}\n.note[data-v-1e7e8ff5] table.table th[align="left"] {\n  text-align: left;\n}\n.note[data-v-1e7e8ff5] table.table th[align="center"] {\n  text-align: center;\n}\n.note[data-v-1e7e8ff5] table.table th[align="right"] {\n  text-align: right;\n}\n',""])},Boyg:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.form-control[data-v-396d57c6] {\n  width: auto;\n}\n",""])},CbGa:function(t,e,n){var s=n("HPB2");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},"D1a/":function(t,e,n){var s=n("WGSt");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},EPOQ:function(t,e,n){"use strict";var s=n("veFt");n.n(s).a},F4tQ:function(t,e,n){"use strict";var s=n("WwcV");n.n(s).a},GYTP:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.slide-fade-enter-active[data-v-764fdbfc] {\n  transition: all .3s ease;\n}\n.slide-fade-leave-active[data-v-764fdbfc] {\n  transition: all .5s cubic-bezier(1.0, 0.5, 0.8, 1.0);\n}\n.slide-fade-enter[data-v-764fdbfc], .slide-fade-leave-to[data-v-764fdbfc] {\n  background-color: #ffef0029;\n}\n",""])},HPB2:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.divider[data-v-01380892] {\n  border-left: 1px solid #ccc;\n}\n",""])},IEyY:function(t,e,n){var s=n("b+/E");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},J82J:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.slide-enter-active[data-v-1a92697e] {\n  transition-duration: 0.3s;\n  transition-timing-function: ease-in;\n}\n.slide-leave-active[data-v-1a92697e] {\n  transition-duration: 0.3s;\n  transition-timing-function: cubic-bezier(0, 1, 0.5, 1);\n}\n.slide-enter-to[data-v-1a92697e], .slide-leave[data-v-1a92697e] {\n  max-height: 1000px;\n  overflow: hidden;\n}\n.slide-enter[data-v-1a92697e], .slide-leave-to[data-v-1a92697e] {\n  overflow: hidden;\n  max-height: 0;\n}\n.text-line-through[data-v-1a92697e] {\n  text-decoration: line-through;\n}\n.btn-collapse[data-v-1a92697e] {\n  font-family: 'Inconsolata', monospace;\n  padding: 0;\n  height: 23px;\n}\n",""])},JRqE:function(t,e,n){var s=n("7HpC");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},L7bX:function(t,e,n){"use strict";var s=n("CbGa");n.n(s).a},"O//m":function(t,e,n){"use strict";var s=n("JRqE");n.n(s).a},P9EP:function(t,e,n){"use strict";var s=n("avtZ");n.n(s).a},RnhZ:function(t,e,n){var s={"./af":"K/tc","./af.js":"K/tc","./ar":"jnO4","./ar-dz":"o1bE","./ar-dz.js":"o1bE","./ar-kw":"Qj4J","./ar-kw.js":"Qj4J","./ar-ly":"HP3h","./ar-ly.js":"HP3h","./ar-ma":"CoRJ","./ar-ma.js":"CoRJ","./ar-sa":"gjCT","./ar-sa.js":"gjCT","./ar-tn":"bYM6","./ar-tn.js":"bYM6","./ar.js":"jnO4","./az":"SFxW","./az.js":"SFxW","./be":"H8ED","./be.js":"H8ED","./bg":"hKrs","./bg.js":"hKrs","./bm":"p/rL","./bm.js":"p/rL","./bn":"kEOa","./bn.js":"kEOa","./bo":"0mo+","./bo.js":"0mo+","./br":"aIdf","./br.js":"aIdf","./bs":"JVSJ","./bs.js":"JVSJ","./ca":"1xZ4","./ca.js":"1xZ4","./cs":"PA2r","./cs.js":"PA2r","./cv":"A+xa","./cv.js":"A+xa","./cy":"l5ep","./cy.js":"l5ep","./da":"DxQv","./da.js":"DxQv","./de":"tGlX","./de-at":"s+uk","./de-at.js":"s+uk","./de-ch":"u3GI","./de-ch.js":"u3GI","./de.js":"tGlX","./dv":"WYrj","./dv.js":"WYrj","./el":"jUeY","./el.js":"jUeY","./en-au":"Dmvi","./en-au.js":"Dmvi","./en-ca":"OIYi","./en-ca.js":"OIYi","./en-gb":"Oaa7","./en-gb.js":"Oaa7","./en-ie":"4dOw","./en-ie.js":"4dOw","./en-il":"czMo","./en-il.js":"czMo","./en-in":"7C5Q","./en-in.js":"7C5Q","./en-nz":"b1Dy","./en-nz.js":"b1Dy","./en-sg":"t+mt","./en-sg.js":"t+mt","./eo":"Zduo","./eo.js":"Zduo","./es":"iYuL","./es-do":"CjzT","./es-do.js":"CjzT","./es-us":"Vclq","./es-us.js":"Vclq","./es.js":"iYuL","./et":"7BjC","./et.js":"7BjC","./eu":"D/JM","./eu.js":"D/JM","./fa":"jfSC","./fa.js":"jfSC","./fi":"gekB","./fi.js":"gekB","./fil":"1ppg","./fil.js":"1ppg","./fo":"ByF4","./fo.js":"ByF4","./fr":"nyYc","./fr-ca":"2fjn","./fr-ca.js":"2fjn","./fr-ch":"Dkky","./fr-ch.js":"Dkky","./fr.js":"nyYc","./fy":"cRix","./fy.js":"cRix","./ga":"USCx","./ga.js":"USCx","./gd":"9rRi","./gd.js":"9rRi","./gl":"iEDd","./gl.js":"iEDd","./gom-deva":"qvJo","./gom-deva.js":"qvJo","./gom-latn":"DKr+","./gom-latn.js":"DKr+","./gu":"4MV3","./gu.js":"4MV3","./he":"x6pH","./he.js":"x6pH","./hi":"3E1r","./hi.js":"3E1r","./hr":"S6ln","./hr.js":"S6ln","./hu":"WxRl","./hu.js":"WxRl","./hy-am":"1rYy","./hy-am.js":"1rYy","./id":"UDhR","./id.js":"UDhR","./is":"BVg3","./is.js":"BVg3","./it":"bpih","./it-ch":"bxKX","./it-ch.js":"bxKX","./it.js":"bpih","./ja":"B55N","./ja.js":"B55N","./jv":"tUCv","./jv.js":"tUCv","./ka":"IBtZ","./ka.js":"IBtZ","./kk":"bXm7","./kk.js":"bXm7","./km":"6B0Y","./km.js":"6B0Y","./kn":"PpIw","./kn.js":"PpIw","./ko":"Ivi+","./ko.js":"Ivi+","./ku":"JCF/","./ku.js":"JCF/","./ky":"lgnt","./ky.js":"lgnt","./lb":"RAwQ","./lb.js":"RAwQ","./lo":"sp3z","./lo.js":"sp3z","./lt":"JvlW","./lt.js":"JvlW","./lv":"uXwI","./lv.js":"uXwI","./me":"KTz0","./me.js":"KTz0","./mi":"aIsn","./mi.js":"aIsn","./mk":"aQkU","./mk.js":"aQkU","./ml":"AvvY","./ml.js":"AvvY","./mn":"lYtQ","./mn.js":"lYtQ","./mr":"Ob0Z","./mr.js":"Ob0Z","./ms":"6+QB","./ms-my":"ZAMP","./ms-my.js":"ZAMP","./ms.js":"6+QB","./mt":"G0Uy","./mt.js":"G0Uy","./my":"honF","./my.js":"honF","./nb":"bOMt","./nb.js":"bOMt","./ne":"OjkT","./ne.js":"OjkT","./nl":"+s0g","./nl-be":"2ykv","./nl-be.js":"2ykv","./nl.js":"+s0g","./nn":"uEye","./nn.js":"uEye","./oc-lnc":"Fnuy","./oc-lnc.js":"Fnuy","./pa-in":"8/+R","./pa-in.js":"8/+R","./pl":"jVdC","./pl.js":"jVdC","./pt":"8mBD","./pt-br":"0tRk","./pt-br.js":"0tRk","./pt.js":"8mBD","./ro":"lyxo","./ro.js":"lyxo","./ru":"lXzo","./ru.js":"lXzo","./sd":"Z4QM","./sd.js":"Z4QM","./se":"//9w","./se.js":"//9w","./si":"7aV9","./si.js":"7aV9","./sk":"e+ae","./sk.js":"e+ae","./sl":"gVVK","./sl.js":"gVVK","./sq":"yPMs","./sq.js":"yPMs","./sr":"zx6S","./sr-cyrl":"E+lV","./sr-cyrl.js":"E+lV","./sr.js":"zx6S","./ss":"Ur1D","./ss.js":"Ur1D","./sv":"X709","./sv.js":"X709","./sw":"dNwA","./sw.js":"dNwA","./ta":"PeUW","./ta.js":"PeUW","./te":"XLvN","./te.js":"XLvN","./tet":"V2x9","./tet.js":"V2x9","./tg":"Oxv6","./tg.js":"Oxv6","./th":"EOgW","./th.js":"EOgW","./tl-ph":"Dzi0","./tl-ph.js":"Dzi0","./tlh":"z3Vd","./tlh.js":"z3Vd","./tr":"DoHr","./tr.js":"DoHr","./tzl":"z1FC","./tzl.js":"z1FC","./tzm":"wQk9","./tzm-latn":"tT3J","./tzm-latn.js":"tT3J","./tzm.js":"wQk9","./ug-cn":"YRex","./ug-cn.js":"YRex","./uk":"raLr","./uk.js":"raLr","./ur":"UpQW","./ur.js":"UpQW","./uz":"Loxo","./uz-latn":"AQ68","./uz-latn.js":"AQ68","./uz.js":"Loxo","./vi":"KSF8","./vi.js":"KSF8","./x-pseudo":"/X5v","./x-pseudo.js":"/X5v","./yo":"fzPg","./yo.js":"fzPg","./zh-cn":"XDpg","./zh-cn.js":"XDpg","./zh-hk":"SatO","./zh-hk.js":"SatO","./zh-mo":"OmwH","./zh-mo.js":"OmwH","./zh-tw":"kOpN","./zh-tw.js":"kOpN"};function a(t){var e=r(t);return n(e)}function r(t){if(!n.o(s,t)){var e=new Error("Cannot find module '"+t+"'");throw e.code="MODULE_NOT_FOUND",e}return s[t]}a.keys=function(){return Object.keys(s)},a.resolve=r,t.exports=a,a.id="RnhZ"},TCw4:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"#main-navigation .router-link-active[data-v-4f12204c] {\n  color: #3490dc;\n}",""])},WGSt:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.bookmark-story[data-v-cdd94d26] {\n  position: relative;\n  top: 2px;\n  text-decoration: none;\n}\n.loader-text[data-v-cdd94d26] {\n  background: #808080;\n  border-radius: 2px;\n}\n.loader-text--1x3[data-v-cdd94d26] {\n  width: 33%;\n}\n.loader-text--2x3[data-v-cdd94d26] {\n  width: 66%;\n}\n",""])},WTCb:function(t,e,n){"use strict";var s=n("k1kw");n.n(s).a},WobJ:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\ntextarea[data-v-c5e2bf76] {\n  outline: 0;\n  background: #f4f4f4;\n  box-shadow: 1px 1px 0 #ccc;\n  border-color: transparent;\n  border-radius: 0;\n  font-family: 'Inconsolata', monospace;\n}\ntextarea[data-v-c5e2bf76]:focus {\n  box-shadow: 1px 1px 0 #ccc;\n  outline: 0;\n  background: #f4f4f4;\n  border-color: transparent;\n}\n.new-tag-group[data-v-c5e2bf76] {\n  font-size: 0;\n}\n.new-tag-group .input-new-tag[data-v-c5e2bf76], .new-tag-group .button-new-tag[data-v-c5e2bf76] {\n  vertical-align: middle;\n  width: 120px;\n}\n#note-create[data-v-c5e2bf76] .el-checkbox-button--small .el-checkbox-button__inner {\n  border-radius: 4px;\n  border: 1px solid #DCDFE6;\n}\n#note-create[data-v-c5e2bf76] .el-checkbox-button.is-checked .el-checkbox-button__inner {\n  border-color: #409EFF;\n}\n",""])},WwcV:function(t,e,n){var s=n("GYTP");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},avtZ:function(t,e,n){var s=n("J82J");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},"b+/E":function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n#users-list[data-v-e8c17f74] .el-table .success-row {\n  background: #f0f9eb;\n}\n@media (max-width: 575px) {\n#users-list[data-v-e8c17f74] .el-dialog {\n    width: 100%;\n}\n}\n",""])},bUC5:function(t,e,n){"use strict";n.r(e);var s=n("XuX8"),a=n.n(s),r=n("jE9Z"),o=n("o0o1"),i=n.n(o),c=n("vDqi"),l=n.n(c),u=n("yDJ3"),d=n.n(u);function m(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}function p(t){return function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){m(r,s,a,o,i,"next",t)}function i(t){m(r,s,a,o,i,"throw",t)}o(void 0)}))}}var f={name:"Authenticate",data:function(){return{user:{email:"",password:""},pollingPasswordFieldInterval:null,errorMessage:null}},methods:{pollingPasswordField:function(){var t=this;this.pollingPasswordFieldInterval=setInterval((function(){var e=document.createEvent("HTMLEvents");e.initEvent("change",!1,!0),t.$refs.pwf.dispatchEvent(e)}),500)},passwordChanged:function(){this.user.password=this.$refs.pwf.value},attemptLogin:function(t){var e=this;return p(i.a.mark((function n(){return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:return t.preventDefault(),n.next=3,l.a.post("/logout");case 3:return n.next=5,l.a.get("/sanctum/csrf-cookie");case 5:return n.next=7,e.login();case 7:case"end":return n.stop()}}),n)})))()},login:function(){var t=this;return p(i.a.mark((function e(){var n,s,a,r,o;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n=422,s=302,a={validateStatus:function(t){return t>=200&&t<300||t===n||t===s}},e.next=5,l.a.post("/login",t.user,a);case 5:if(r=e.sent,(o=r.status)!==n){e.next=11;break}t.showsErrors(d()(r,"data.errors")),e.next=16;break;case 11:if(o!==s&&204!==o){e.next=16;break}return clearInterval(t.pollingPasswordFieldInterval),t.$message.success("Login successful"),e.next=16,t.$router.push(Pt);case 16:case"end":return e.stop()}}),e)})))()},showsErrors:function(t){var e="";Object.keys(t).forEach((function(n){t[n].forEach((function(t){e+=t+"<br/>"}))})),this.$message({type:"error",dangerouslyUseHTMLString:!0,message:e})}},mounted:function(){this.pollingPasswordField()}},v=n("KHd+"),h=Object(v.a)(f,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"pt-4",attrs:{id:"view-authenticate"}},[n("div",{staticClass:"row justify-content-center"},[n("div",{staticClass:"col-md-8"},[n("div",{staticClass:"card shadow"},[n("div",{staticClass:"card-header"},[t._v(t._s(t.$I18n.trans("auth.login")))]),t._v(" "),n("div",{staticClass:"card-body"},[n("form",{attrs:{method:"POST",model:t.user},on:{submit:function(e){return t.attemptLogin(e)}}},[n("div",{staticClass:"form-group row"},[n("label",{staticClass:"col-md-4 col-form-label text-md-right",attrs:{for:"email"}},[t._v(t._s(t.$I18n.trans("auth.email")))]),t._v(" "),n("div",{staticClass:"col-md-6"},[n("input",{directives:[{name:"model",rawName:"v-model",value:t.user.email,expression:"user.email"}],staticClass:"form-control",attrs:{id:"email",inputmode:"email",type:"email",name:"email",autocomplete:"email",required:""},domProps:{value:t.user.email},on:{input:function(e){e.target.composing||t.$set(t.user,"email",e.target.value)}}})])]),t._v(" "),n("div",{staticClass:"form-group row"},[n("label",{staticClass:"col-md-4 col-form-label text-md-right",attrs:{for:"password"}},[t._v(t._s(t.$I18n.trans("auth.password")))]),t._v(" "),n("div",{staticClass:"col-md-6"},[n("input",{directives:[{name:"model",rawName:"v-model",value:t.user.password,expression:"user.password"}],ref:"pwf",staticClass:"form-control",attrs:{id:"password",name:"password",type:"password",autocomplete:"current-password",required:""},domProps:{value:t.user.password},on:{change:t.passwordChanged,input:function(e){e.target.composing||t.$set(t.user,"password",e.target.value)}}})])]),t._v(" "),n("div",{staticClass:"form-group row"},[n("div",{staticClass:"col-md-6 offset-md-4"},[n("div",{staticClass:"form-check"},[n("input",{staticClass:"form-check-input",attrs:{type:"checkbox",name:"remember",id:"remember",checked:""}}),t._v(" "),n("label",{staticClass:"form-check-label",attrs:{for:"remember"}},[t._v(t._s(t.$I18n.trans("auth.remember_me")))])])])]),t._v(" "),n("div",{staticClass:"form-group row mb-0"},[n("div",{staticClass:"col-md-8 offset-md-4"},[n("button",{staticClass:"btn btn-primary",attrs:{type:"submit",disabled:!t.user.email&&!t.user.password}},[t._v(t._s(t.$I18n.trans("auth.login")))])])])])])])])])])}),[],!1,null,null,null).exports,g={name:"NotFound",props:{message:{type:String,required:!1,default:"We could not find what you where looking for!"},linksTo:{type:Array,required:!1}},data:function(){return{dashboardRoute:Pt}}},b=Object(v.a)(g,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"pt-4",attrs:{id:"NotFound"}},[n("h1",{staticClass:"text-center text-capitalize"},[t._v("404 - page not found")]),t._v(" "),n("h5",{staticClass:"text-center text-black-50 mt-4"},[t._v(t._s(t.message))]),t._v(" "),n("div",{staticClass:"px-5 pt-3 text-center"},[n("router-link",{staticClass:"btn btn-lg btn-link",attrs:{to:t.dashboardRoute}},[t._v("Home")]),t._v(" "),t._l(t.linksTo,(function(e){return n("router-link",{key:e.route.name,staticClass:"btn btn-lg btn-link",attrs:{to:e.route}},[t._v(t._s(t.route.linkText))])}))],2)])}),[],!1,null,"98208f0c",null).exports,_=n("wd/R"),x=n.n(_);function k(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var y={name:"Weather",data:function(){return{loading:!0,weather:{icon:"",tempFeelsLike:"",humidity:"",description:"",sunrise:"",sunset:""}}},methods:{fetchWeather:function(){var t,e=this;return(t=i.a.mark((function t(){var n,s,a,r,o;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:n="openweather--currentweather",a=localStorage.getItem(n);try{s=JSON.parse(a)}catch(t){s=null}if(s&&s.data&&!(s.saved+6e5<Date.now())){t.next=12;break}return s={data:{},saved:Date.now()},t.next=7,fetch("http://api.openweathermap.org/data/2.5/weather?q=Funchal,PT&appid=fe66d26d5b01501290be42a4809bbc4e&units=metric&lang=pt");case 7:return r=t.sent,t.next=10,r.json();case 10:s.data=t.sent,localStorage.setItem(n,JSON.stringify(s));case 12:o=d()(s,"data.weather[0].icon"),e.weather.icon="http://openweathermap.org/img/wn/".concat(o,".png"),e.weather.tempFeelsLike=d()(s,"data.main.feels_like"),e.weather.temp=d()(s,"data.main.temp"),e.weather.description=d()(s,"data.weather[0].description"),e.weather.humidity=d()(s,"data.main.humidity"),e.weather.sunrise=d()(s,"data.sys.sunrise"),e.weather.sunset=d()(s,"data.sys.sunset"),setTimeout((function(){return e.loading=!1}),500);case 21:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){k(r,s,a,o,i,"next",t)}function i(t){k(r,s,a,o,i,"throw",t)}o(void 0)}))})()}},filters:{capitalize:function(t){if(t){var e=t.toString();return e.charAt(0).toUpperCase()+e.slice(1)}},localTimeFromUnixTimestamp:function(t){if(t)return x.a.unix(t).format("HH:mm")}},created:function(){this.fetchWeather(),setInterval(this.fetchWeather,6e5)}},w=(n("L7bX"),Object(v.a)(y,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],attrs:{id:"weather"}},[n("div",{staticClass:"card mb-3 shadow"},[n("div",{staticClass:"card-body d-flex flex-wrap"},[n("div",{staticClass:"align-self-center"},[n("a",{attrs:{href:"https://openweathermap.org/city/2267827",target:"_blank"}},[n("img",{attrs:{src:t.weather.icon,alt:t.weather.description}})])]),t._v(" "),n("div",{staticClass:"divider align-items-stretch mx-2"}),t._v(" "),n("div",{staticClass:"align-self-center ml-2"},[t._v("\n        "+t._s(t._f("capitalize")(t.weather.description))+"\n      ")]),t._v(" "),n("div",{staticClass:"divider align-items-stretch mx-2"}),t._v(" "),n("div",{staticClass:"align-self-center"},[n("div",{staticClass:"d-flex"},[n("div",{staticClass:"align-items-stretch align-self-center"},[t._v("Temps:")]),t._v(" "),n("div",{staticClass:"d-flex flex-column ml-1"},[n("div",{attrs:{title:"Feels like"}},[t._v(t._s(t.weather.tempFeelsLike)+" ℃")]),t._v(" "),n("div",{attrs:{title:"Temperature"}},[n("small",[t._v(t._s(t.weather.temp)+" ℃")])])])])]),t._v(" "),n("div",{staticClass:"divider align-items-stretch mx-2"}),t._v(" "),n("div",{staticClass:"align-self-center"},[n("div",{staticClass:"d-flex"},[n("div",{staticClass:"flex-column"},[n("div",[t._v("Humidade")]),t._v(" "),n("div",{staticClass:"text-center"},[t._v(t._s(t.weather.humidity)+"%")])])])]),t._v(" "),n("div",{staticClass:"divider align-items-stretch mx-2"}),t._v(" "),n("div",{staticClass:"align-self-center"},[n("div",{staticClass:"d-flex"},[n("div",{staticClass:"flex-column ml-1"},[n("div",{attrs:{title:"Sunrise"}},[t._v(t._s(t._f("localTimeFromUnixTimestamp")(t.weather.sunrise)))]),t._v(" "),n("div",{attrs:{title:"Sunset"}},[t._v(t._s(t._f("localTimeFromUnixTimestamp")(t.weather.sunset)))])])])])])])])}),[],!1,null,"01380892",null).exports);function C(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var j={name:"SystemInfo",data:function(){return{loading:!0,memory:{free:"",used:"",total:""},nQueueWorkersRunning:0,up:"",upSince:"",systemInfoLoop:null}},created:function(){this.fetchSystemInfo(),this.systemInfoLoop=setInterval(this.fetchSystemInfo,1e4)},methods:{fetchSystemInfo:function(){var t,e=this;return(t=i.a.mark((function t(){var n,s;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,l.a.get("/api/system-info");case 2:n=t.sent,(s=d()(n,"data.data",null))&&(e.memory.free=d()(s,"memory.free"),e.memory.used=d()(s,"memory.used"),e.memory.total=d()(s,"memory.total"),e.nQueueWorkersRunning=d()(s,"nQueueWorkersRunning"),e.up=d()(s,"up"),e.upSince=d()(s,"upSince")),setTimeout((function(){return e.loading=!1}),500);case 6:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){C(r,s,a,o,i,"next",t)}function i(t){C(r,s,a,o,i,"throw",t)}o(void 0)}))})()}},filters:{capitalize:function(t){if(t)return(t=t.toString()).charAt(0).toUpperCase()+t.slice(1)}}},I=(n("F4tQ"),Object(v.a)(j,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"card mb-3 shadow"},[n("div",{staticClass:"card-header"},[t._v(t._s(t.$I18n.trans("system.info")))]),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"card-body"},[n("transition",{attrs:{name:"slide-fade",mode:"out-in"}},[n("p",{key:t.up,staticClass:"alert-info"},[t._v("\n        "+t._s(t._f("capitalize")(t.up))+" "+t._s(t.$I18n.trans("system.since"))+" "+t._s(t.upSince)+"\n      ")])]),t._v(" "),n("p",{class:{"alert-success":t.nQueueWorkersRunning,"alert-danger":!t.nQueueWorkersRunning}},[t._v("\n      "+t._s(t.$I18n.trans("system.number_queue_workers"))+": "+t._s(t.nQueueWorkersRunning)+"\n    ")]),t._v(" "),n("div",{staticClass:"pt-4"},[n("table",{staticClass:"table table-sm"},[n("caption",[t._v(t._s(t.$I18n.trans("system.since")))]),t._v(" "),n("thead",[n("tr",[n("th",[t._v(t._s(t.$I18n.trans("system.used")))]),t._v(" "),n("th",[t._v(t._s(t.$I18n.trans("system.free")))]),t._v(" "),n("th",[t._v(t._s(t.$I18n.trans("system.total")))])])]),t._v(" "),n("tbody",[n("transition",{attrs:{name:"slide-fade",mode:"out-in"}},[n("tr",{key:t.memory.used},[n("td",[t.memory.used?t._e():n("span",[t._v(" ")]),t._v(t._s(t.memory.used))]),t._v(" "),n("td",[t._v(t._s(t.memory.free))]),t._v(" "),n("td",[t._v(t._s(t.memory.total))])])])],1)])])],1)])}),[],!1,null,"764fdbfc",null).exports);function $(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var T={name:"Inspire",data:function(){return{inspire:null,fallbackLines:["The hardest part of building software is deciding what to build.","Everything not specified will be specified by the programmer, on the fly.","Design is also specification.","Everything that is public-facing defines the significant part of the business logic.","The open office plan is the best way to destroy productivity.","Don’t multitask. Focus on one task at a time.","There is work time and there is personal time. Don’t mix the two.","Everyone writes bad code except for you. /s"]}},computed:{fallback:function(){return this.fallbackLines[Math.floor(Math.random()*this.fallbackLines.length)]}},created:function(){this.fetchInspire()},methods:{fetchInspire:function(){var t,e=this;return(t=i.a.mark((function t(){var n;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,l.a.get("/api/inspire");case 2:n=t.sent,e.inspire=d()(n,"data.data");case 4:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){$(r,s,a,o,i,"next",t)}function i(t){$(r,s,a,o,i,"throw",t)}o(void 0)}))})()}}},E=Object(v.a)(T,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"card mb-3 shadow"},[n("div",{staticClass:"card-body"},[n("small",[t.inspire?t._e():n("span",[t._v(t._s(t.fallback))]),t._v(t._s(t.inspire))])])])}),[],!1,null,null,null).exports;function S(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var N={name:"NextHolidays",data:function(){return{loading:!0,nextHolidays:[{name:"",date:"",type:"",description:""},{name:"",date:"",type:"",description:""},{name:"",date:"",type:"",description:""}]}},methods:{fetchNextHolidays:function(){var t,e=this;return(t=i.a.mark((function t(){var n,s;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,l.a.get("/api/next-holidays");case 2:return n=t.sent,(s=d()(n,"data.data",[])).length&&(e.nextHolidays=s.map((function(t){return{name:t.name,description:t.description,date:t.date.iso,type:Array.isArray(t.type)?t.type[0]:""}})),e.loading=!1),t.abrupt("return",!0);case 6:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){S(r,s,a,o,i,"next",t)}function i(t){S(r,s,a,o,i,"throw",t)}o(void 0)}))})()}},created:function(){this.fetchNextHolidays()}},P={name:"Dashboard",components:{Weather:w,SystemInfo:I,NextHolidays:Object(v.a)(N,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"card shadow mb-3"},[n("div",{staticClass:"card-header"},[t._v(t._s(t.$I18n.trans("holidays.next_holidays")))]),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"card-body"},t._l(t.nextHolidays,(function(e){return n("div",[n("p",[n("span",{staticClass:"badge badge-dark"},[t._v("#")]),t._v("\n        "+t._s(e.name)+",\n        "+t._s(t._f("diffForHumans")(e.date))+"\n        at "+t._s(e.date)+"\n        "),n("small",[t._v(t._s(e.type))])]),t._v(" "),n("p",[e.description?t._e():n("span",[t._v(" ")]),t._v(" "),n("small",[t._v(t._s(e.description))])])])})),0)])}),[],!1,null,null,null).exports,Inspire:E},data:function(){return{}},beforeRouteLeave:function(t,e,n){clearInterval(this.$refs.refSystemInfo.systemInfoLoop),n()}},O=Object(v.a)(P,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{staticClass:"pt-4",attrs:{id:"view-dashboard"}},[e("div",{staticClass:"row justify-content-center"},[e("div",{staticClass:"col-sm-6 no-gutter-xs"},[e("inspire"),this._v(" "),e("system-info",{ref:"refSystemInfo"})],1),this._v(" "),e("div",{staticClass:"col-sm-6 no-gutter-xs"},[e("weather"),this._v(" "),e("next-holidays")],1)])])}),[],!1,null,null,null).exports,B={name:"Navigation",data:function(){return{notesRoute:Ot,notesCreateRoute:Bt}},methods:{activeSubmenu:function(t){return this.$route.name===t?"btn-primary":"btn-secondary"}}},D=Object(v.a)(B,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100",class:t.activeSubmenu("Notes"),attrs:{to:t.notesRoute}},[t._v(t._s(t.$I18n.trans("notes.notes")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100",class:t.activeSubmenu("NotesCreate"),attrs:{to:t.notesCreateRoute}},[t._v(t._s(t.$I18n.trans("notes.new_note")))])],1)])])])}),[],!1,null,null,null).exports,U=n("9/5/"),R={name:"AlgoliaSearch",data:function(){return{searchQuery:""}},watch:{searchQuery:function(){(0===this.searchQuery.length||this.searchQuery.length>1)&&this.submit()}},methods:{submit:n.n(U)()((function(){this.$emit("inputData",this.searchQuery)}),400),clearSearch:function(){this.searchQuery="",this.submit()}}};n("maHV");function F(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var z={name:"NotesList",props:["searchQuery"],data:function(){return{notesShowRoute:Dt,loading:!0,notes:[]}},created:function(){var t=this;this.fetchNotes(this.searchQuery).then((function(e){t.loading=!1}))},watch:{searchQuery:function(){var t=this;this.loading=!0,this.fetchNotes(this.searchQuery).then((function(e){t.loading=!1}))}},methods:{fetchNotes:function(t){var e,n=this;return(e=i.a.mark((function e(){var s;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,l.a.get("/api/notes?query=".concat(t));case 2:return s=e.sent,n.notes=d()(s,"data.data",[]),e.abrupt("return",!0);case 5:case"end":return e.stop()}}),e)})),function(){var t=this,n=arguments;return new Promise((function(s,a){var r=e.apply(t,n);function o(t){F(r,s,a,o,i,"next",t)}function i(t){F(r,s,a,o,i,"throw",t)}o(void 0)}))})()},close:function(){this.$emit("inputData","")}}},H={name:"Notes",data:function(){return{searchQuery:""}},components:{Navigation:D,AlgoliaSearch:Object(v.a)(R,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"col-12 no-gutter-xs",attrs:{id:"algolia-search"}},[n("div",{staticClass:"justify-content-center form-inline d-flex"},[n("label",{staticClass:"sr-only",attrs:{for:"query"}},[t._v(t._s(t.$I18n.trans("notes.search")))]),t._v(" "),n("input",{directives:[{name:"model",rawName:"v-model",value:t.searchQuery,expression:"searchQuery"}],staticClass:"form-control flex-grow-1",attrs:{type:"text",id:"query",name:"query",placeholder:t.$I18n.trans("notes.min_2_chars")},domProps:{value:t.searchQuery},on:{keyup:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"esc",27,e.key,["Esc","Escape"])?null:t.clearSearch(e)},input:function(e){e.target.composing||(t.searchQuery=e.target.value)}}}),t._v("\n     \n    "),t._m(0)])])}),[function(){var t=this.$createElement,e=this._self._c||t;return e("picture",[e("source",{attrs:{srcset:"search-by-algolia.svg",media:"(min-width: 800px)"}}),this._v(" "),e("img",{staticStyle:{"max-height":"30px"},attrs:{src:"algolia-blue-mark.svg",alt:"search by algolia"}})])}],!1,null,"396d57c6",null).exports,NotesList:Object(v.a)(z,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"notes-list"}},[n("ul",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"list-group list-group-flush"},[t.notes.length?t._e():n("el-alert",{attrs:{title:t.$I18n.trans("notes.alert_failed_to_find_records"),type:"info",center:"","close-text":t.$I18n.trans("notes.alert_close_text")},on:{close:t.close}}),t._v(" "),t._l(t.notes,(function(e){return n("li",{staticClass:"list-group-item list-group-item-action p-2 p-sm-3"},[n("router-link",{staticClass:"text-secondary",attrs:{to:{name:t.notesShowRoute.name,params:{id:e.id}}}},[n("div",{staticClass:"d-flex w-100 justify-content-between"},[n("h5",{staticClass:"mb-1"},[t._v(t._s(e.title))]),t._v(" "),n("small",{staticClass:"d-none d-sm-block"},[t._v(t._s(t._f("diffForHumans")(e.updated_at)))])])]),t._v(" "),n("div",t._l(e.tags,(function(e){return n("el-tag",{key:e.id,staticClass:"mr-1 mb-1",attrs:{size:"small"}},[t._v(t._s(e.name))])})),1)],1)}))],2)])}),[],!1,null,null,null).exports},methods:{updateMessage:function(t){this.searchQuery=t}}},A=Object(v.a)(H,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"notes"}},[e("navigation"),this._v(" "),e("div",{staticClass:"row justify-content-center"},[e("algolia-search",{on:{inputData:this.updateMessage}})],1),this._v(" "),e("div",{staticClass:"row"},[e("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[e("notes-list",{attrs:{"search-query":this.searchQuery},on:{inputData:this.updateMessage}})],1)])],1)}),[],!1,null,null,null).exports,M={methods:{textareaCharInserter:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"  ";t.preventDefault();var n=this.note.body,s=t.target.selectionStart,a=n.slice(0,s),r=n.slice(s);this.note.body="".concat(a).concat(e).concat(r),t.target.value=this.note.body,t.target.selectionEnd=t.target.selectionStart=s+1}}};function L(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}function Q(t){return function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){L(r,s,a,o,i,"next",t)}function i(t){L(r,s,a,o,i,"throw",t)}o(void 0)}))}}var V={name:"NotesCreate",components:{Navigation:D},mixins:[M],data:function(){return{newTagInputVisible:!1,newTag:"",errors:[],note:{body:""},tags:[],selectedTags:[],inputEl:null}},methods:{insertCodeBlockBackTicks:function(){if(this.inputEl||(this.inputEl=document.querySelector("#note-create textarea")),!document.execCommand("insertText",!1,"```")&&"function"==typeof this.inputEl.setRangeText){var t=this.inputEl.selectionStart;this.inputEl.setRangeText("```"),this.inputEl.selectionStart=this.inputEl.selectionEnd=t+"```".length;var e=document.createEvent("UIEvent");e.initEvent("input",!0,!1),this.inputEl.dispatchEvent(e)}},addNewTag:function(){var t=this;return Q(i.a.mark((function e(){var n,s,a,r;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(!t.newTag){e.next=15;break}return t.errors=[],e.prev=2,e.next=5,l.a.post("/api/notetagcreate",{tag:t.newTag});case 5:n=e.sent,s=d()(n,"data.data.success"),a=d()(n,"data.data.id"),s&&a&&(t.tags.push({id:a,name:t.newTag}),t.newTagInputVisible=!1,t.newTag=""),e.next=15;break;case 11:e.prev=11,e.t0=e.catch(2),r=d()(e.t0,"response.data.errors.tag[0]"),t.errors.push({field:"tagCreate",text:r});case 15:case"end":return e.stop()}}),e,null,[[2,11]])})))()},showInput:function(){var t=this;this.newTagInputVisible=!0,this.$nextTick((function(e){t.$refs.saveTagInput.$refs.input.focus()}))},fetchTags:function(){var t=this;return Q(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,l.a.get("/api/notes/tags");case 2:n=e.sent,t.tags=d()(n,"data.data",[]);case 4:case"end":return e.stop()}}),e)})))()},submitForm:function(t){var e=this;return Q(i.a.mark((function n(){var s,a,r;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t.preventDefault(),!e.newTagInputVisible){n.next=3;break}return n.abrupt("return",!1);case 3:if(e.errors=[],e.note.body){n.next=8;break}return e.errors.push({field:"body",text:"Note text is empty."}),e.note.body=null,n.abrupt("return");case 8:return n.prev=8,n.next=11,l.a.post("/api/notecreate",{body:e.note.body,tags:e.selectedTags});case 11:s=n.sent,a=d()(s,"data.data.id"),r=d()(s,"data.data.success"),a&&r?(e.note.body="",e.selectedTags=[],e.$message({type:"success",message:e.$I18n.trans("notes.confirmation_success_note_create"),center:!0}),setTimeout((function(){return e.$router.push(Object.assign({},Dt,{params:{id:a}}))}),400)):e.errors.push({field:"na",text:"Failed to create the note."}),n.next=20;break;case 17:n.prev=17,n.t0=n.catch(8),e.errors.push({field:"submit",text:n.t0.message});case 20:case"end":return n.stop()}}),n,null,[[8,17]])})))()}},created:function(){this.fetchTags()}},Y=(n("EPOQ"),Object(v.a)(V,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"notes-create"}},[n("navigation"),t._v(" "),n("div",{staticClass:"row justify-content-center"},[n("div",{staticClass:"col-sm no-gutter-xs"},[n("div",{staticClass:"card shadow"},[t.errors.length?n("div",{staticClass:"p-1"},t._l(t.errors,(function(t){return n("el-alert",{key:t.field,attrs:{title:t.text,type:"error"}})})),1):t._e(),t._v(" "),n("form",{attrs:{method:"post"},on:{submit:t.submitForm}},[n("div",{staticClass:"textarea-container"},[n("textarea",{directives:[{name:"model",rawName:"v-model",value:t.note.body,expression:"note.body"}],staticClass:"form-control form-text m-0",attrs:{placeholder:t.$I18n.trans("notes.placeholder_note_title"),name:"body",cols:"100",rows:"15"},domProps:{value:t.note.body},on:{keydown:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"tab",9,e.key,"Tab")?null:(e.preventDefault(),t.textareaCharInserter(e))},input:function(e){e.target.composing||t.$set(t.note,"body",e.target.value)}}})]),t._v(" "),n("div",{staticClass:"d-flex justify-content-between"},[n("small",{staticClass:"form-text text-muted ml-3"},[t._m(0)]),t._v(" "),n("el-button",{attrs:{circle:"",size:"small",type:"info"},on:{click:t.insertCodeBlockBackTicks}},[t._v("```")])],1),t._v(" "),n("div",{staticClass:"mt-3 ml-3"},[n("p",[t._v(t._s(t.$I18n.trans("notes.tags"))+":")]),t._v(" "),n("el-checkbox-group",{staticClass:"d-inline-block",attrs:{size:"small"},model:{value:t.selectedTags,callback:function(e){t.selectedTags=e},expression:"selectedTags"}},t._l(t.tags,(function(e){return n("el-checkbox-button",{key:e.id,staticClass:"mr-1",attrs:{label:e.id}},[t._v(t._s(e.name))])})),1),t._v(" "),n("div",{staticClass:"d-inline-block new-tag-group"},[t.newTagInputVisible?n("el-input",{ref:"saveTagInput",staticClass:"input-new-tag",attrs:{size:"small"},on:{blur:t.addNewTag},nativeOn:{keyup:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"enter",13,e.key,"Enter")?null:t.addNewTag(e)}},model:{value:t.newTag,callback:function(e){t.newTag=e},expression:"newTag"}}):n("el-button",{staticClass:"button-new-tag",attrs:{type:"info",plain:"",size:"small"},on:{click:t.showInput}},[t._v(t._s(t.$I18n.trans("notes.new_tag")))])],1)],1),t._v(" "),n("div",{staticClass:"px-3 pb-3 pt-3"},[n("button",{staticClass:"btn btn-primary",attrs:{type:"submit"}},[t._v(t._s(t.$I18n.trans("common.add")))])])])])])])],1)}),[function(){var t=this.$createElement;return(this._self._c||t)("a",{staticClass:"text-muted",attrs:{href:"https://commonmark.org/help/",tabindex:"-1",target:"_CommonMark"}},[this._v(this._s(this.$I18n.trans("notes.commonmark")))])}],!1,null,"c5e2bf76",null).exports);n("WxZA");function q(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}function W(t){return function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){q(r,s,a,o,i,"next",t)}function i(t){q(r,s,a,o,i,"throw",t)}o(void 0)}))}}var J={name:"NotesShow",components:{Navigation:D},props:["id"],data:function(){return{loading:!0,note:{},notesUpdateRoute:Ut}},created:function(){this.fetchNote(this.id)},methods:{fetchNote:function(t){var e=this;return W(i.a.mark((function n(){var s;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t){n.next=2;break}return n.abrupt("return",!1);case 2:return n.prev=2,n.next=5,l.a.get("/api/notes/".concat(t,"?html=1"));case 5:s=n.sent,n.next=14;break;case 8:if(n.prev=8,n.t0=n.catch(2),404!==n.t0.response.status){n.next=14;break}return n.next=13,e.$router.push(Object.assign({},Qt,{params:{message:"Note could not be found.",linksTo:[{route:Ot,linkText:"Notes"}]}}));case 13:return n.abrupt("return");case 14:e.note=d()(s,"data.data"),setTimeout((function(){return e.loading=!1}),400);case 16:case"end":return n.stop()}}),n,null,[[2,8]])})))()},confirmDelete:function(t){var e=this;return W(i.a.mark((function n(){var s,a,r,o;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:s=e.$I18n.trans("notes.confirmation_ask_delete",{id:t}),a=e.$I18n.trans("notes.confirmation_success_deleted",{id:t}),r=e.$I18n.trans("notes.delete"),o=e.$I18n.trans("notes.cancel"),e.$confirm(s,"Warning",{confirmButtonText:r,cancelButtonText:o,type:"warning",center:!0}).then((function(){e.deleteNote(t).then((function(t){t&&(e.$message({type:"success",message:a,center:!0}),setTimeout((function(){return e.$router.push(Ot)}),400))})).catch((function(n){e.$message({type:"warning",message:e.$I18n.trans("notes.failed_delete",{id:t}),center:!0})}))})).catch((function(){e.$message({type:"info",message:e.$I18n.trans("notes.delete_canceled",{id:t}),center:!0})}));case 5:case"end":return n.stop()}}),n)})))()},deleteNote:function(t){return W(i.a.mark((function e(){var n,s;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,l.a.delete("/api/notedestroy/".concat(t),{_method:"delete"});case 2:if(n=e.sent,!(s=d()(n,"data.data.success",!1))){e.next=6;break}return e.abrupt("return",s);case 6:throw"Delete failed";case 7:case"end":return e.stop()}}),e)})))()}}},X=(n("WTCb"),Object(v.a)(J,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"note"},[n("navigation"),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"row justify-content-center"},[n("div",{staticClass:"col-12 no-gutter-xs"},[t.note.id?t._e():n("div",{staticClass:"mt-3 "},[n("p",{staticClass:"loader-text pb-2 pt-3"},[t._v(" ")]),t._v(" "),n("p",{staticClass:"loader-text"},[t._v(" ")]),t._v(" "),n("p",{staticClass:"loader-text"},[t._v(" ")]),t._v(" "),n("p",{staticClass:"loader-text loader-text--1x3 d-inline-block"},[t._v("   ")])]),t._v(" "),n("div",{staticClass:"mt-3",domProps:{innerHTML:t._s(t.note.body)}}),t._v(" "),n("div",t._l(t.note.tags,(function(e){return n("el-tag",{key:e.id,staticClass:"mr-1 mb-1",attrs:{size:"big"}},[t._v(t._s(e.name))])})),1)])]),t._v(" "),n("div",{staticClass:"row"},[n("div",{staticClass:"col-12"},[n("div",{staticClass:"mt-3"},[t.note.id?n("router-link",{staticClass:"btn btn-primary",attrs:{to:Object.assign({},t.notesUpdateRoute,{params:{id:t.note.id}})}},[t._v(t._s(t.$I18n.trans("notes.edit")))]):t._e(),t._v(" "),n("button",{staticClass:"btn btn-danger ml-3",on:{click:function(e){return t.confirmDelete(t.note.id)}}},[t._v(t._s(t.$I18n.trans("notes.delete")))])],1)])])],1)}),[],!1,null,"1e7e8ff5",null).exports);function K(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}function Z(t){return function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){K(r,s,a,o,i,"next",t)}function i(t){K(r,s,a,o,i,"throw",t)}o(void 0)}))}}var G={name:"NotesUpdate",components:{Navigation:D},mixins:[M],props:["id"],data:function(){return{loading:!0,newTagInputVisible:!1,newTag:"",errors:[],note:{},tags:[{id:1,name:""}],selectedTags:[],inputEl:null}},methods:{insertCodeBlockBackTicks:function(){if(this.inputEl||(this.inputEl=document.querySelector("#note-update textarea")),!document.execCommand("insertText",!1,"```")&&"function"==typeof this.inputEl.setRangeText){var t=this.inputEl.selectionStart;this.inputEl.setRangeText("```"),this.inputEl.selectionStart=this.inputEl.selectionEnd=t+"```".length;var e=document.createEvent("UIEvent");e.initEvent("input",!0,!1),this.inputEl.dispatchEvent(e)}},addNewTag:function(){var t=this;return Z(i.a.mark((function e(){var n,s,a,r;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(!t.newTag){e.next=15;break}return t.errors=[],e.prev=2,e.next=5,l.a.post("/api/notetagcreate",{tag:t.newTag});case 5:n=e.sent,s=d()(n,"data.data.success"),a=d()(n,"data.data.id"),s&&a&&(t.tags.push({id:a,name:t.newTag}),t.newTagInputVisible=!1,t.newTag=""),e.next=15;break;case 11:e.prev=11,e.t0=e.catch(2),r=d()(e.t0,"response.data.errors.tag[0]"),t.errors.push({field:"tagCreate",text:r});case 15:case"end":return e.stop()}}),e,null,[[2,11]])})))()},showInput:function(){var t=this;this.newTagInputVisible=!0,this.$nextTick((function(e){t.$refs.saveTagInput.$refs.input.focus()}))},fetchNote:function(t){var e=this;return Z(i.a.mark((function n(){var s;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t){n.next=2;break}return n.abrupt("return",!1);case 2:return n.next=4,l.a.get("/api/notes/".concat(t,"?html=0"));case 4:s=n.sent,e.note=d()(s,"data.data",[]),e.selectedTags=e.note.tags.map((function(t){return t.id})),setTimeout((function(){return e.loading=!1}),400);case 8:case"end":return n.stop()}}),n)})))()},fetchTags:function(){var t=this;return Z(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,l.a.get("/api/notes/tags");case 2:n=e.sent,t.tags=d()(n,"data.data",[]);case 4:case"end":return e.stop()}}),e)})))()},submitForm:function(t){var e=this;return Z(i.a.mark((function n(){var s;return i.a.wrap((function(n){for(;;)switch(n.prev=n.next){case 0:if(t.preventDefault(),!e.newTagInputVisible){n.next=3;break}return n.abrupt("return",!1);case 3:if(e.errors=[],e.note.body){n.next=7;break}return e.errors.push({field:"body",text:"Note text is empty."}),n.abrupt("return");case 7:return n.prev=7,n.next=10,l.a.put("/api/noteupdate/".concat(e.note.id),{body:e.note.body,tags:e.selectedTags,_method:"put"});case 10:s=n.sent,d()(s,"data.data.success")?(e.selectedTags=[],e.$message({type:"success",message:e.$I18n.trans("notes.confirmation_success_note_updated"),center:!0}),setTimeout((function(){return e.$router.push(Object.assign({},Dt,{params:{id:e.note.id}}))}),400)):e.errors.push({field:"na",text:"Failed to update note."}),n.next=17;break;case 14:n.prev=14,n.t0=n.catch(7),e.errors.push({field:"submit",text:n.t0.message});case 17:case"end":return n.stop()}}),n,null,[[7,14]])})))()}},created:function(){var t=this;this.fetchTags().then((function(){t.fetchNote(t.id)}))}},tt=(n("O//m"),Object(v.a)(G,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"notes-update"}},[n("navigation"),t._v(" "),n("div",{staticClass:"row justify-content-center"},[n("div",{staticClass:"col-sm no-gutter-xs"},[n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"card shadow"},[t.errors.length?n("div",{staticClass:"p-1"},t._l(t.errors,(function(t){return n("el-alert",{key:t.field,attrs:{title:t.text,type:"error"}})})),1):t._e(),t._v(" "),n("form",{attrs:{method:"post"},on:{submit:t.submitForm}},[n("div",{staticClass:"textarea-container"},[n("textarea",{directives:[{name:"model",rawName:"v-model",value:t.note.body,expression:"note.body"}],staticClass:"form-control form-text m-0",attrs:{placeholder:t.$I18n.trans("notes.placeholder_note_title"),name:"body",cols:"100",rows:"15"},domProps:{value:t.note.body},on:{keydown:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"tab",9,e.key,"Tab")?null:(e.preventDefault(),t.textareaCharInserter(e))},input:function(e){e.target.composing||t.$set(t.note,"body",e.target.value)}}})]),t._v(" "),n("div",{staticClass:"d-flex justify-content-between"},[n("small",{staticClass:"form-text text-muted ml-3"},[t._m(0)]),t._v(" "),n("el-button",{attrs:{circle:"",size:"small",type:"info"},on:{click:t.insertCodeBlockBackTicks}},[t._v("```")])],1),t._v(" "),n("div",{staticClass:"mt-3 ml-3"},[n("p",[t._v("Tags:")]),t._v(" "),n("el-checkbox-group",{staticClass:"d-inline-block",attrs:{size:"small"},model:{value:t.selectedTags,callback:function(e){t.selectedTags=e},expression:"selectedTags"}},t._l(t.tags,(function(e){return n("el-checkbox-button",{key:e.id,staticClass:"mr-1",attrs:{label:e.id}},[t._v(t._s(e.name))])})),1),t._v(" "),n("div",{staticClass:"d-inline-block new-tag-group"},[t.newTagInputVisible?n("el-input",{ref:"saveTagInput",staticClass:"input-new-tag",attrs:{size:"small"},on:{blur:t.addNewTag},nativeOn:{keyup:function(e){return!e.type.indexOf("key")&&t._k(e.keyCode,"enter",13,e.key,"Enter")?null:t.addNewTag(e)}},model:{value:t.newTag,callback:function(e){t.newTag=e},expression:"newTag"}}):n("el-button",{staticClass:"button-new-tag",attrs:{type:"info",plain:"",size:"small"},on:{click:t.showInput}},[t._v(t._s(t.$I18n.trans("notes.new_tag")))])],1)],1),t._v(" "),n("div",{staticClass:"px-3 pb-3 pt-3"},[n("button",{staticClass:"btn btn-primary",attrs:{type:"submit"}},[t._v(t._s(t.$I18n.trans("common.update")))])])])])])])],1)}),[function(){var t=this.$createElement;return(this._self._c||t)("a",{staticClass:"text-muted",attrs:{href:"https://commonmark.org/help/",tabindex:"-1",target:"_CommonMark"}},[this._v(this._s(this.$I18n.trans("notes.commonmark")))])}],!1,null,"4b83dcc8",null).exports),et={name:"Navigation",props:{numberOfBookmarks:{type:null|Number,required:!0}},data:function(){return{hackerNewsTopPostsRoute:Rt,hackerNewsBestPostsRoute:Ft,hackerNewsBookmarkedPostsRoute:zt}}},nt=Object(v.a)(et,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.hackerNewsTopPostsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.top")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.hackerNewsBestPostsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.best")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.hackerNewsBookmarkedPostsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.bookmarks"))+" "),n("span",{staticClass:"badge badge-light"},[t._v(t._s(t.numberOfBookmarks))])])],1)])])])}),[],!1,null,null,null).exports,st={name:"HackerNews",components:{Navigation:nt},data:function(){return{numberOfBookmarks:null}},methods:{updateNumberOfBookmarks:function(t){this.numberOfBookmarks=t}}},at=Object(v.a)(st,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"hacker-news"}},[e("navigation",{attrs:{"number-of-bookmarks":this.numberOfBookmarks}}),this._v(" "),e("router-view",{on:{nBookmarksChangedEvent:this.updateNumberOfBookmarks}})],1)}),[],!1,null,null,null).exports;function rt(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var ot={data:function(){return{nBookmarks:null,notification:null}},methods:{bookmarkPost:function(t){var e,n=this;return(e=i.a.mark((function e(){var s,a,r,o;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return t.status.saving=!0,s={postId:t.id},(a=!t.status.bookmarked)||(s._method="delete"),e.next=6,l.a.post("/api/bookmarks",s);case 6:r=e.sent,(o=d()(r,"data.data.success",!1))&&(t.status.bookmarked=!t.status.bookmarked,n.nBookmarks+=a?1:-1),n.notifyUserOfBookmarkStatusChange(o,a,t);case 10:case"end":return e.stop()}}),e)})),function(){var t=this,n=arguments;return new Promise((function(s,a){var r=e.apply(t,n);function o(t){rt(r,s,a,o,i,"next",t)}function i(t){rt(r,s,a,o,i,"throw",t)}o(void 0)}))})()},notifyUserOfBookmarkStatusChange:function(t,e,n){var s={message:this.$I18n.trans("hackernews.add_failure"),type:"error"};t&&(s=e?{message:this.$I18n.trans("hackernews.added_to_bookmarks",{title:n.title}),type:"success"}:{message:this.$I18n.trans("hackernews.remove_from_bookmarks",{title:n.title}),type:"warning"}),this.notification&&!this.notification.closed&&this.notification.close(),this.notification=this.$message(s),setTimeout((function(){return n.status.saving=!1}),400)}}},it=n("Wcq6"),ct=n.n(it),lt=(n("Zs65"),ct.a.initializeApp({databaseURL:"hacker-news.firebaseio.com"}).database().ref("/v0"));function ut(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}function dt(t){return function(t){if(Array.isArray(t))return mt(t)}(t)||function(t){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t))return Array.from(t)}(t)||function(t,e){if(!t)return;if("string"==typeof t)return mt(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return mt(t,e)}(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function mt(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,s=new Array(e);n<e;n++)s[n]=t[n];return s}var pt={name:"HNPosts",mixins:[ot],props:{idList:{type:Array,default:[],required:!0}},data:function(){return{loading:!0,posts:[],nBookmarks:null,firstBatchPosition:10,hackerNewsPostRoute:Ht}},methods:{fetchItems:function(t){var e=this,n=this.firstBatchPosition;t.length<n&&(n=Math.floor(t.length/2));var s=t.slice(0,n),a=t.slice(n);Promise.all(s.map((function(t){return e.fetchItem(t)}))).then((function(t){e.posts=t.filter(e.isStoryValid),e.attachBookmarked(),e.loading=!1})),this.$nextTick((function(){Promise.all(a.map((function(t){return e.fetchItem(t)}))).then((function(t){var n;(n=e.posts).push.apply(n,dt(t.filter(e.isStoryValid))),e.attachBookmarked()}))}))},isStoryValid:function(t){return"story"===t.type&&!!t.id},fetchItem:function(t){return new Promise((function(e,n){lt.child("item/".concat(t)).on("value",(function(t){var n=t.val(),s={id:d()(n,"id"),title:d()(n,"title"),score:d()(n,"score"),time:x.a.unix(d()(n,"time",0)).format("YYYY-MM-DD HH:mm:ss"),url:d()(n,"url"),type:d()(n,"type"),nComments:d()(n,"descendants"),kids:d()(n,"kids",[]),status:{bookmarked:!1,saving:!1}};e(s)})),lt.onDisconnect((function(t,e){this.$alert("Firebase connection lost","Lost connection",{confirmButtonText:"OK"}),console.log(t,e)}))}))},attachBookmarked:function(){var t,e=this;return(t=i.a.mark((function t(){var n,s;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,l.a.get("/api/bookmarks");case 2:n=t.sent,s=d()(n,"data.data",[]),e.nBookmarks=s.length,e.posts=e.posts.map((function(t){return t.status.bookmarked=s.includes(t.id),t}));case 6:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){ut(r,s,a,o,i,"next",t)}function i(t){ut(r,s,a,o,i,"throw",t)}o(void 0)}))})()}},watch:{idList:function(t,e){var n=this;this.loading=!0,setTimeout((function(){return n.fetchItems(t)}),400)},nBookmarks:function(t,e){this.$emit("nBookmarksChangedEvent",t)}}};n("0P3G");function ft(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var vt={name:"HackerNewsList",components:{Navigation:nt,HnPosts:Object(v.a)(pt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("ul",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticClass:"list-group",attrs:{id:"hn-posts"}},[t._l(Array(t.firstBatchPosition).fill(null),(function(e){return t.loading?n("li",{staticClass:"list-group-item list-group-item-action flex-column align-items-start"},[t._m(0,!0),t._v(" "),t._m(1,!0)]):t._e()})),t._v(" "),t._l(t.posts,(function(e,s){return n("li",{key:e.id,staticClass:"list-group-item list-group-item-action flex-column align-items-start",attrs:{"data-index":s+1}},[n("div",{staticClass:"d-flex w-100 justify-content-between"},[n("span",[n("router-link",{staticClass:"text-body",attrs:{to:Object.assign({},t.hackerNewsPostRoute,{params:{id:e.id}})}},[t._v(t._s(e.title))]),t._v(" "),e.url?n("a",{staticClass:"text-body",attrs:{href:e.url,target:"_blank"}},[n("small",{staticClass:"text-muted"},[t._v("("+t._s(t._f("domainFromUrl")(e.url))+")")])]):t._e()],1),t._v(" "),n("span",{directives:[{name:"loading",rawName:"v-loading",value:e.status.saving,expression:"post.status.saving"}],staticClass:"d-block d-md-none"},[n("span",{staticClass:"bookmark-story pointer text-primary",on:{click:function(n){return t.bookmarkPost(e)}}},[t._v(t._s(e.status.bookmarked?"⚫":"⚪️")+"️")])]),t._v(" "),n("span",{staticClass:"badge d-none d-md-block",attrs:{title:e.time}},[t._v(t._s(t._f("diffForHumans")(e.time)))])]),t._v(" "),n("div",{directives:[{name:"loading",rawName:"v-loading",value:e.status.saving,expression:"post.status.saving"}],staticClass:"d-none d-md-block"},[n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.points",{points:e.score||""})))]),t._v("\n      |\n      "),n("small",[t._v(t._s(t.$I18n.trans("hackernews.comments",{comments:e.nComments||""})))]),t._v("\n      |\n      "),n("span",{staticClass:"bookmark-story pointer text-primary",on:{click:function(n){return t.bookmarkPost(e)}}},[t._v(t._s(e.status.bookmarked?"⚫":"⚪️"))])])])}))],2)}),[function(){var t=this.$createElement,e=this._self._c||t;return e("div",{staticClass:"d-flex w-100 justify-content-between"},[e("span",{staticClass:"loader-text loader-text-top d-block w-100"},[this._v(" ")])])},function(){var t=this.$createElement,e=this._self._c||t;return e("div",{staticClass:"d-none d-md-block"},[e("span",{staticClass:"loader-text loader-text-bottom loader-text--1x3 d-inline-block"},[e("small",[this._v(" ")])])])}],!1,null,"a624b0ec",null).exports},props:{type:{type:String,validator:function(t){return-1!==["top","best","bookmarks"].indexOf(t)},default:"top",required:!0}},data:function(){return{idList:[],limit:100,types:{HackerNewsBookmarks:"bookmarks"},numberOfBookmarks:null}},methods:{fetchStories:function(){this.type===this.types.HackerNewsBookmarks?this.fetchIdsBookmarkedFromBackend():this.fetchIdsFromFirebase(this.type)},fetchIdsFromFirebase:function(t){var e=this;lt.child("".concat(t,"stories")).limitToFirst(this.limit).once("value",(function(t,n){var s=t.val();e.idList=Array.isArray(s)?s:[]}))},fetchIdsBookmarkedFromBackend:function(){var t,e=this;return(t=i.a.mark((function t(){var n;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,l.a.get("/api/bookmarks");case 2:n=t.sent,e.idList=d()(n,"data.data",[]);case 4:case"end":return t.stop()}}),t)})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){ft(r,s,a,o,i,"next",t)}function i(t){ft(r,s,a,o,i,"throw",t)}o(void 0)}))})()},updateNumberOfBookmarks:function(t){this.numberOfBookmarks=t}},created:function(){this.fetchStories()},watch:{$route:function(t,e){this.fetchStories()}}},ht=Object(v.a)(vt,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"hacker-news-list"}},[e("div",{staticClass:"row"},[e("div",{staticClass:"col-12 no-gutter-xs"},[e("hn-posts",this._g({attrs:{"id-list":this.idList}},this.$listeners))],1)])])}),[],!1,null,null,null).exports,gt={name:"HnComment",props:{comment:{type:Object,required:!0},handleCollapseToggle:{type:Function,required:!0}},data:function(){return{isShow:!0}}};n("P9EP");function bt(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var _t={name:"HackerNewsPost",mixins:[ot],components:{Navigation:nt,HnComment:Object(v.a)(gt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"hn-comment card mt-3"},[n("div",{staticClass:"card-body px-2 px-md-3 shadow"},[n("h6",{staticClass:"card-subtitle bg-light text-muted d-flex justify-content-between align-items-center",class:{"opacity-5 text-line-through":t.comment.deleted}},[t.comment.id?n("span",{staticClass:"align-middle"},[t._v(t._s(t.$I18n.trans("hackernews.by",{by:t.comment.by||""})))]):t._e(),t._v(" "),n("small",{staticClass:"text-muted",attrs:{title:t.comment.time}},[t._v(t._s(t._f("diffForHumans")(t.comment.time)))]),t._v(" "),t.comment.id?n("a",{staticClass:"btn btn-sm pointer btn-collapse",attrs:{role:"button","data-toggle":"collapse"},on:{click:function(e){return t.handleCollapseToggle(t.comment)}}},[n("b",[t._v("["),t.comment.collapsed?n("span",[t._v(t._s(t.comment.kids.length))]):n("span",[t._v("-")]),t._v("]\n        ")])]):t._e()]),t._v(" "),n("transition",{attrs:{name:"slide"}},[t.comment.collapsed?t._e():n("div",{staticClass:"mt-2"},[n("p",{domProps:{innerHTML:t._s(t.comment.text)}}),t._v(" "),t._l(t.comment.comments,(function(e){return t.comment.comments.length?n("hn-comment",{key:e.id,attrs:{comment:e,"handle-collapse-toggle":t.handleCollapseToggle}}):t._e()}))],2)])],1)])}),[],!1,null,"1a92697e",null).exports},props:{id:{type:Number|String,required:!0}},data:function(){return{loading:!0,loadingComments:!0,collapsedCommentsKey:"collapsedCommentsHackerNewsPost",collapsedComments:[],post:{id:null,by:null,title:null,text:null,score:null,time:null,url:null,type:null,nComments:null,status:{bookmarked:null,saving:null},kids:[],comments:[{id:null,by:null,kids:[],parent:null,text:null,time:null,comments:[]}]},numberOfBookmarks:null}},methods:{fetchPost:function(t){var e=this;t&&(this.fetchCollapsedComments(),lt.child("item/".concat(t)).once("value",(function(t){var n=t.val();e.post.id=d()(n,"id"),e.post.by=d()(n,"by"),e.post.title=d()(n,"title"),e.post.text=d()(n,"text"),e.post.score=d()(n,"score"),e.post.time=x.a.unix(d()(n,"time",0)).format("YYYY-MM-DD HH:mm:ss"),e.post.url=d()(n,"url"),e.post.type=d()(n,"type"),e.post.nComments=d()(n,"descendants"),e.post.bookmarked=null,e.post.kids=d()(n,"kids",[]),e.post.comments=[],e.setBookmarkStatus(e.post),e.fetchComments(e.post),e.loading=!1})))},fetchComments:function(t){var e=this;t.kids.length&&t.kids.map((function(n){e.fetchComment(n).then((function(n){var s=d()(n,"kids",[]),a=d()(n,"id"),r={id:a,deleted:d()(n,"deleted"),by:d()(n,"by"),kids:s,text:d()(n,"text"),time:x.a.unix(d()(n,"time",0)).format("YYYY-MM-DD HH:mm:ss"),comments:[],collapsed:e.collapsedComments.includes(a)};e.fetchComments(r),t.comments.push(r),e.loadingComments=!1}))}))},fetchComment:function(t){return new Promise((function(e,n){lt.child("item/".concat(t)).once("value",(function(t){var n=t.val();e(n)}))}))},setBookmarkStatus:function(t){var e,n=this;return(e=i.a.mark((function e(){var s,a;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,l.a.get("/api/bookmarks");case 2:s=e.sent,a=d()(s,"data.data",[]),n.numberOfBookmarks=a.length,t.status.bookmarked=a.includes(t.id);case 6:case"end":return e.stop()}}),e)})),function(){var t=this,n=arguments;return new Promise((function(s,a){var r=e.apply(t,n);function o(t){bt(r,s,a,o,i,"next",t)}function i(t){bt(r,s,a,o,i,"throw",t)}o(void 0)}))})()},fetchCollapsedComments:function(){var t,e=localStorage.getItem(this.collapsedCommentsKey);try{t=JSON.parse(e)}catch(t){console.error(t)}t=Array.isArray(t)?t:[],this.collapsedComments=t},setCollapsedComments:function(){var t=JSON.stringify(this.collapsedComments);localStorage.setItem(this.collapsedCommentsKey,t)},handleCollapseToggle:function(t){t.collapsed=!t.collapsed,t.collapsed?this.collapsedComments.push(t.id):this.collapsedComments=this.collapsedComments.filter((function(e){return e!==t.id})),this.setCollapsedComments()}},created:function(){this.fetchPost(this.id)}},xt=(n("zTjN"),Object(v.a)(_t,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"hacker-news-post"}},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}]},[t.loading?n("p",{staticClass:"lead loader-text d-block loader-text--2x3"},[t._v(" ")]):t._e(),t._v(" "),t.post.title?n("p",{staticClass:"lead"},[t._v("\n          "+t._s(t.post.title)+"\n          "),t.post.url?n("a",{staticClass:"text-body",attrs:{href:t.post.url,target:"_blank"}},[n("small",{staticClass:"text-muted"},[t._v("("+t._s(t._f("domainFromUrl")(t.post.url))+") [↗]")])]):t._e()]):t._e(),t._v(" "),t.post.text?n("p",{domProps:{innerHTML:t._s(t.post.text)}}):t._e(),t._v(" "),t.loading?n("div",{staticClass:"loader-text d-block loader-text--1x3"},[t._v(" ")]):t._e(),t._v(" "),t.post.title?n("div",{directives:[{name:"loading",rawName:"v-loading",value:t.post.status.saving,expression:"post.status.saving"}]},[n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.points",{points:t.post.score})))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.comments",{comments:t.post.nComments})))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("small",{staticClass:"text-muted",attrs:{title:t.post.time}},[t._v(t._s(t._f("diffForHumans")(t.post.time)))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("small",{staticClass:"text-muted"},[t._v(t._s(t.$I18n.trans("hackernews.by",{by:t.post.by})))]),t._v(" "),n("span",{staticClass:"text-muted"},[t._v("|")]),t._v(" "),n("span",{staticClass:"bookmark-story pointer text-primary",on:{click:function(e){return t.bookmarkPost(t.post)}}},[t._v(t._s(t.post.status.bookmarked?"⚫":"⚪️")+"️")])]):t._e()])])]),t._v(" "),t.post.comments.length?n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 no-gutter-xs"},[t.loadingComments?t._e():n("div",t._l(t.post.comments,(function(e){return n("hn-comment",{key:e.id,attrs:{comment:e,"handle-collapse-toggle":t.handleCollapseToggle}})})),1)])]):t._e()])}),[],!1,null,"cdd94d26",null).exports),kt={name:"Navigation",data:function(){return{usersRoute:Lt,usersCreateRoute:Mt}}},yt={name:"Users",components:{Navigation:Object(v.a)(kt,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"navigation"}},[n("div",{staticClass:"row justify-content-center top-submenu"},[n("div",{staticClass:"col-12 no-gutter-xs"},[n("div",{staticClass:"btn-group d-flex mb-2"},[n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.usersRoute}},[t._v(t._s(t.$I18n.trans("users.users")))]),t._v(" "),n("router-link",{staticClass:"btn btn-group-sm w-100 inner-nav btn-secondary",attrs:{to:t.usersCreateRoute}},[t._v(t._s(t.$I18n.trans("users.new-user")))])],1)])])])}),[],!1,null,null,null).exports}},wt=Object(v.a)(yt,(function(){var t=this.$createElement,e=this._self._c||t;return e("div",{attrs:{id:"users"}},[e("navigation"),this._v(" "),e("router-view")],1)}),[],!1,null,null,null).exports;function Ct(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}function jt(t){return function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){Ct(r,s,a,o,i,"next",t)}function i(t){Ct(r,s,a,o,i,"throw",t)}o(void 0)}))}}var It={name:"Users",data:function(){return{loading:!0,users:[],userToEdit:{},currentUserId:null,dialogVisible:!1,dialogFormRules:{name:[{required:!0,message:this.$I18n.trans("users.name_cannot_be_empty"),trigger:"blur"},{min:2,max:255,message:this.$I18n.trans("users.min_2_max_255"),trigger:"blur"}]}}},methods:{fetchUsers:function(){var t=this;return jt(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,l.a.get("/api/users");case 2:n=e.sent,t.users=d()(n,"data.data.users",[]),t.currentUserId=d()(n,"data.data.currentUserId",[]),t.loading=!1;case 6:case"end":return e.stop()}}),e)})))()},highlightCurrentUser:function(t){var e=t.row;t.rowIndex;return e.id===this.currentUserId?"table-success":null},formatDate:function(t){return this.$options.filters.diffForHumans(t)},formatCreatedAt:function(t,e){return this.formatDate(t.createdAt)},formatUpdatedAt:function(t,e){return this.formatDate(t.updatedAt)},closeDialog:function(t){this.dialogVisible=!1,this.$refs[t].resetFields()},editDialogShow:function(t){this.userToEdit={id:t.id,name:t.name,email:t.email},this.dialogVisible=!0},editDialogSubmit:function(t){var e=this;this.$refs[t].validate((function(t){return!!t&&(e.updateUser(),!0)}))},updateUser:function(){var t=this;return jt(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,e.next=3,l.a.put("/api/users/update",{id:t.userToEdit.id,name:t.userToEdit.name});case 3:n=e.sent,e.next=11;break;case 6:return e.prev=6,e.t0=e.catch(0),console.error(e.t0),t.$message.error(t.$I18n.trans("users.failed_to_update")),e.abrupt("return");case 11:t.notifyActionSuccess(n);case 12:case"end":return e.stop()}}),e,null,[[0,6]])})))()},notifyActionSuccess:function(t){var e=this,n=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"users.user_updated",s=arguments.length>2&&void 0!==arguments[2]?arguments[2]:"users.failed_to_update",a=d()(t,"status"),r=d()(t,"data.data.success",!1);if(200===a&&r){this.$message.success(this.$I18n.trans(n));var o=this.users.find((function(t){return t.id===e.userToEdit.id}));o.name=this.userToEdit.name.trim(),this.dialogVisible=!1}else this.$message.error(this.$I18n.trans(s))},deleteUser:function(){var t=this;return jt(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,t.askUserToConfirmDelete();case 2:if(!e.sent){e.next=18;break}return e.prev=4,e.next=7,l.a.delete("/api/users/destroy",{data:{id:t.userToEdit.id}});case 7:n=e.sent,console.log(n),e.next=16;break;case 11:return e.prev=11,e.t0=e.catch(4),console.error(e.t0),t.$message.error(t.$I18n.trans("users.failed_to_delete")),e.abrupt("return");case 16:t.notifyActionSuccess(n,"users.delete_success","users.failed_to_delete"),t.fetchUsers();case 18:case"end":return e.stop()}}),e,null,[[4,11]])})))()},askUserToConfirmDelete:function(){var t=this;return jt(i.a.mark((function e(){var n;return i.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.prev=0,e.next=3,t.$confirm(t.$I18n.trans("users.confirm_delete_request",{name:t.userToEdit.name}),t.$I18n.trans("common.please_confirm"),{confirmButtonText:t.$I18n.trans("users.delete"),cancelButtonText:t.$I18n.trans("common.cancel"),type:"warning"});case 3:e.t0=e.sent,n="confirm"===e.t0,e.next=10;break;case 7:e.prev=7,e.t1=e.catch(0),n=!1;case 10:return e.abrupt("return",n);case 11:case"end":return e.stop()}}),e,null,[[0,7]])})))()}},created:function(){this.fetchUsers()}},$t=(n("4101"),Object(v.a)(It,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"users-list"}},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[n("el-table",{directives:[{name:"loading",rawName:"v-loading",value:t.loading,expression:"loading"}],staticStyle:{width:"100%"},attrs:{data:t.users,"row-class-name":t.highlightCurrentUser,"empty-text":t.$I18n.trans("users.failed_to_obtain_user_list")}},[n("el-table-column",{attrs:{prop:"id",label:"#"}}),t._v(" "),n("el-table-column",{attrs:{prop:"name",label:t.$I18n.trans("users.name")}}),t._v(" "),n("el-table-column",{attrs:{prop:"email",label:t.$I18n.trans("users.email")}}),t._v(" "),n("el-table-column",{attrs:{prop:"createdAt",label:t.$I18n.trans("users.created_at"),formatter:t.formatCreatedAt}}),t._v(" "),n("el-table-column",{attrs:{prop:"updatedAt",label:t.$I18n.trans("users.updated_at"),formatter:t.formatUpdatedAt}}),t._v(" "),n("el-table-column",{attrs:{fixed:"right",align:"right",width:"60",prop:"updatedAt"},scopedSlots:t._u([{key:"default",fn:function(e){return[n("el-button",{attrs:{type:"text",size:"small"},on:{click:function(n){return t.editDialogShow(e.row)}}},[t._v(t._s(t.$I18n.trans("users.edit")))])]}}])})],1)],1)]),t._v(" "),n("el-dialog",{attrs:{title:t.$I18n.trans("users.user_edit",{id:t.userToEdit.id||""}),visible:t.dialogVisible},on:{"update:visible":function(e){t.dialogVisible=e}}},[n("el-form",{ref:"formRef",attrs:{model:t.userToEdit,rules:t.dialogFormRules},on:{submit:t.editDialogSubmit},nativeOn:{submit:function(t){t.preventDefault()}}},[n("el-form-item",{attrs:{label:t.$I18n.trans("users.name"),prop:"name"}},[n("el-input",{attrs:{placeholder:t.$I18n.trans("users.name_cannot_be_empty")},model:{value:t.userToEdit.name,callback:function(e){t.$set(t.userToEdit,"name",e)},expression:"userToEdit.name"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.email")}},[n("el-input",{attrs:{type:"email",value:t.userToEdit.email,disabled:""}})],1)],1),t._v(" "),n("span",{staticClass:"dialog-footer",attrs:{slot:"footer"},slot:"footer"},[n("el-button",{on:{click:function(e){return t.closeDialog("formRef")}}},[t._v(t._s(t.$I18n.trans("common.cancel")))]),t._v(" "),n("el-button",{attrs:{type:"danger",disabled:t.userToEdit.id===t.currentUserId},on:{click:function(e){return t.deleteUser()}}},[t._v(t._s(t.$I18n.trans("users.delete")))]),t._v(" "),n("el-button",{attrs:{type:"primary"},on:{click:function(e){return t.editDialogSubmit("formRef")}}},[t._v(t._s(t.$I18n.trans("common.confirm")))])],1)],1)],1)}),[],!1,null,"e8c17f74",null).exports);function Tt(t,e,n,s,a,r,o){try{var i=t[r](o),c=i.value}catch(t){return void n(t)}i.done?e(c):Promise.resolve(c).then(s,a)}var Et={name:"UsersCreate",data:function(){return{rules:{name:[{required:!0,message:this.$I18n.trans("users.name_cannot_be_empty"),trigger:"blur"},{min:2,max:255,message:this.$I18n.trans("users.min_2_max_255"),trigger:"blur"}],email:[{required:!0,type:"email",message:this.$I18n.trans("users.please_set_valid_email"),trigger:"blur"}],password:[{required:!0,message:this.$I18n.trans("users.please_set_valid_password"),trigger:"blur"},{min:12,max:100,message:this.$I18n.trans("users.please_set_valid_password"),trigger:"blur"}]},user:{name:"",email:"",password:""}}},methods:{submitForm:function(t){var e=this;this.$refs[t].validate((function(t){return!!t&&(e.createUser(),!0)}))},createUser:function(){var t,e=this;return(t=i.a.mark((function t(){var n;return i.a.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.prev=0,t.next=3,l.a.post("/api/users/create",e.user);case 3:n=t.sent,t.next=9;break;case 6:t.prev=6,t.t0=t.catch(0),console.error(t.t0);case 9:e.notifyActionSuccess(n);case 10:case"end":return t.stop()}}),t,null,[[0,6]])})),function(){var e=this,n=arguments;return new Promise((function(s,a){var r=t.apply(e,n);function o(t){Tt(r,s,a,o,i,"next",t)}function i(t){Tt(r,s,a,o,i,"throw",t)}o(void 0)}))})()},notifyActionSuccess:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"users.user_create",n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:"users.failed_to_create",s=d()(t,"status"),a=d()(t,"data.data.success",!1);200===s&&a?(this.$message.success(this.$I18n.trans(e)),this.$router.push(Lt)):this.$message.error(this.$I18n.trans(n))}}},St=Object(v.a)(Et,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"users-create"}},[n("div",{staticClass:"row"},[n("div",{staticClass:"col-12 mt-3 no-gutter-xs"},[n("el-form",{ref:"refCreateUserForm",attrs:{model:t.user,rules:t.rules}},[n("el-form-item",{attrs:{label:t.$I18n.trans("users.name"),prop:"name"}},[n("el-input",{model:{value:t.user.name,callback:function(e){t.$set(t.user,"name",e)},expression:"user.name"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.email"),prop:"email"}},[n("el-input",{attrs:{type:"email",autocomplete:"off"},model:{value:t.user.email,callback:function(e){t.$set(t.user,"email",e)},expression:"user.email"}})],1),t._v(" "),n("el-form-item",{attrs:{label:t.$I18n.trans("users.password"),prop:"password"}},[n("el-input",{attrs:{type:"password",autocomplete:"off"},model:{value:t.user.password,callback:function(e){t.$set(t.user,"password",e)},expression:"user.password"}})],1),t._v(" "),n("el-form-item",[n("el-button",{attrs:{type:"primary"},on:{click:function(e){return t.createUser("refCreateUserForm")}}},[t._v(t._s(t.$I18n.trans("users.create")))])],1)],1)],1)])])}),[],!1,null,null,null).exports;a.a.use(r.a);var Nt={path:"/authenticate",name:"Authenticate",component:h},Pt={path:"/dashboard",alias:"/",name:"Dashboard",component:O},Ot={path:"/notes",name:"Notes",component:A},Bt={path:"/notes/new",name:"NotesCreate",component:Y},Dt={path:"/notes/:id",name:"NotesShow",component:X,props:!0},Ut={path:"/notes/edit/:id",name:"NotesUpdate",component:tt,props:!0},Rt={path:"",alias:"top",name:"HackerNewsTopPostsRoute",component:ht,props:{type:"top"}},Ft={path:"best",name:"HackerNewsBestPostsRoute",component:ht,props:{type:"best"}},zt={path:"bookmarks",name:"HackerNewsBookmarkedPostsRoute",component:ht,props:{type:"bookmarks"}},Ht={path:":id",name:"HackerNewsPost",component:xt,props:!0},At={path:"/hn",component:at,children:[Rt,Ft,zt,Ht]},Mt={path:"new",name:"UsersCreate",component:St},Lt={path:"/users",component:wt,children:[{path:"",name:"UsersList",component:$t},Mt]},Qt={path:"/404",name:"NotFound",component:b,props:!0},Vt=new r.a({routes:[Nt,Pt,Ot,Bt,Dt,Ut,At,Lt,Qt,{path:"*",redirect:"/404"}]}),Yt=n("XJYT"),qt=n.n(Yt);n("D66Q");function Wt(t){return(Wt="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function Jt(t,e){return function(t){if(Array.isArray(t))return t}(t)||function(t,e){if("undefined"==typeof Symbol||!(Symbol.iterator in Object(t)))return;var n=[],s=!0,a=!1,r=void 0;try{for(var o,i=t[Symbol.iterator]();!(s=(o=i.next()).done)&&(n.push(o.value),!e||n.length!==e);s=!0);}catch(t){a=!0,r=t}finally{try{s||null==i.return||i.return()}finally{if(a)throw r}}return n}(t,e)||function(t,e){if(!t)return;if("string"==typeof t)return Xt(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return Xt(t,e)}(t,e)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function Xt(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,s=new Array(e);n<e;n++)s[n]=t[n];return s}function Kt(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function Zt(t,e){for(var n=0;n<e.length;n++){var s=e[n];s.enumerable=s.enumerable||!1,s.configurable=!0,"value"in s&&(s.writable=!0),Object.defineProperty(t,s.key,s)}}a.a.filter("diffForHumans",(function(t){if(t)return x()(t,"YYYY-MM-DD hh:mm:ss").fromNow()})),a.a.filter("domainFromUrl",(function(t){if(t)return new URL(t).hostname.replace("www.","")}));var Gt=function(){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:"translations";Kt(this,t),this.key=e}var e,n,s;return e=t,(n=[{key:"trans",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{};return this._replace(this._extract(t),e)}},{key:"trans_choice",value:function(t){var e,n=this,s=arguments.length>1&&void 0!==arguments[1]?arguments[1]:1,a=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},r=this._extract(t,"|").split("|");return r.some((function(t){return e=n._match(t,s)})),e=(e=e||(s>1?r[1]:r[0])).replace(/\[.*?\]|\{.*?\}/,""),this._replace(e,a)}},{key:"_match",value:function(t,e){var n=t.match(/^[\{\[]([^\[\]\{\}]*)[\}\]](.*)/);if(n){if(n[1].includes(",")){var s=Jt(n[1].split(",",2),2),a=s[0],r=s[1];if("*"===r&&e>=a)return n[2];if("*"===a&&e<=r)return n[2];if(e>=a&&e<=r)return n[2]}return n[1]==e?n[2]:null}}},{key:"_replace",value:function(t,e){if("object"===Wt(t))return t;for(var n in e)t=t.toString().replace(":".concat(n),e[n]).replace(":".concat(n.toUpperCase()),e[n].toString().toUpperCase()).replace(":".concat(n.charAt(0).toUpperCase()).concat(n.slice(1)),e[n].toString().charAt(0).toUpperCase()+e[n].toString().slice(1));return t.toString().trim()}},{key:"_extract",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:null,n=t.toString().split("::"),s=n.pop().toString().split(".");return n.length>0&&(n[0]+="::"),n.concat(s).reduce((function(n,s){return n[s]||e||t}),window[this.key])}}])&&Zt(e.prototype,n),s&&Zt(e,s),t}(),te=n("D/Jt"),ee={name:"MainNavigation",data:function(){return{notesRoute:Ot,dashboardRoute:Pt,hackerNewsRoute:At,usersRoute:Lt,showMenu:!0,menuCollapsed:!0,appRoutes:[]}},methods:{logout:function(t){var e=this;t.preventDefault(),l.a.post("/logout",{}).then((function(){e.$router.push(Nt),e.showMenu=!1}))},toggleMenuVisibility:function(t){this.showMenu=Nt.name!==t.name}},watch:{$route:function(t,e){this.toggleMenuVisibility(t)}},mounted:function(){this.toggleMenuVisibility(this.$route)}},ne=(n("2LfK"),Object(v.a)(ee,(function(){var t=this,e=t.$createElement,n=t._self._c||e;return t.showMenu?n("nav",{staticClass:"navbar navbar-expand-md navbar-light navbar-laravel",attrs:{id:"main-navigation"}},[n("div",{staticClass:"container"},[n("router-link",{staticClass:"navbar-brand",attrs:{exact:"",to:t.dashboardRoute}},[t._v("Zetetic Elench")]),t._v(" "),n("button",{staticClass:"navbar-toggler",attrs:{type:"button"},on:{click:function(e){t.menuCollapsed=!t.menuCollapsed}}},[n("span",{staticClass:"navbar-toggler-icon"})]),t._v(" "),n("div",{staticClass:"navbar-collapse",class:{collapse:t.menuCollapsed}},[n("ul",{staticClass:"navbar-nav mr-auto"},[n("li",{staticClass:"nav-item"},[n("router-link",{staticClass:"nav-link text-right text-sm-left",attrs:{exact:"",to:t.notesRoute}},[t._v(t._s(t.$I18n.trans("notes.notes")))])],1),t._v(" "),n("li",{staticClass:"nav-item"},[n("router-link",{staticClass:"nav-link text-right text-sm-left",attrs:{to:t.hackerNewsRoute}},[t._v(t._s(t.$I18n.trans("hackernews.hackernews")))])],1),t._v(" "),n("li",{staticClass:"nav-item"},[n("router-link",{staticClass:"nav-link text-right text-sm-left",attrs:{to:t.usersRoute}},[t._v(t._s(t.$I18n.trans("users.users")))])],1)]),t._v(" "),n("ul",{staticClass:"navbar-nav ml-auto"},[n("li",{staticClass:"nav-item"},[n("a",{staticClass:"nav-link text-right text-sm-left",attrs:{href:"#"},on:{click:function(e){return t.logout(e)}}},[t._v(t._s(t.$I18n.trans("users.logout")))])])])])],1)]):t._e()}),[],!1,null,"4f12204c",null).exports);a.a.component("main-navigation",ne),a.a.use(qt.a),a.a.use(te.a),a.a.prototype.$I18n=new Gt,l.a.defaults.headers.common["X-Requested-With"]="XMLHttpRequest",l.a.defaults.withCredentials=!0,l.a.interceptors.response.use(null,(function(t){if(t.response)return 401!==t.response.status&&419!==t.response.status||se.$router.push(Nt),Promise.reject(t);console.error("Unknown error: ".concat(t))}));var se=new a.a({router:Vt,el:"#app"})},fysm:function(t,e,n){(t.exports=n("I1BE")(!1)).push([t.i,"\n.bookmark-story[data-v-a624b0ec] {\n  position: relative;\n  top: 2px;\n  text-decoration: none;\n}\n.loader-text[data-v-a624b0ec] {\n  background: #808080;\n  border-radius: 2px;\n}\n.loader-text--1x3[data-v-a624b0ec] {\n  width: 33%;\n}\n.loader-text-top[data-v-a624b0ec] {\n  height: 22px;\n  margin-bottom: 6px;\n}\n.loader-text-bottom[data-v-a624b0ec] {\n  height: 18px;\n}\n",""])},iEU0:function(t,e,n){var s=n("Boyg");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},k1kw:function(t,e,n){var s=n("7Pbq");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},lLLT:function(t,e,n){var s=n("TCw4");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},maHV:function(t,e,n){"use strict";var s=n("iEU0");n.n(s).a},pyCd:function(t,e){},veFt:function(t,e,n){var s=n("WobJ");"string"==typeof s&&(s=[[t.i,s,""]]);var a={hmr:!0,transform:void 0,insertInto:void 0};n("aET+")(s,a);s.locals&&(t.exports=s.locals)},zTjN:function(t,e,n){"use strict";var s=n("D1a/");n.n(s).a}},[[0,1,2]]]);
