+++ public/mix-manifest.json
@@ -1,6 +1,6 @@
 {
-    "/js/app.js": "/js/app.js?id=45945ee1676fcf7b09d2",
-    "/css/app.css": "/css/app.css?id=c05d6b33fc719d664d47",
+    "/js/app.js": "/js/app.js?id=824f282c7364565fb55b",
+    "/css/app.css": "/css/app.css?id=b7e183f5706765b6bc71",
     "/js/manifest.js": "/js/manifest.js?id=3c768977c2574a34506e",
     "/js/vendor.js": "/js/vendor.js?id=0f6a845a57326e140dca"
 }
