+++ resources/js/users/components/Navigation.vue
@@ -5,13 +5,11 @@
         <div class="btn-group d-flex mb-2">
           <router-link
             :to="usersRoute"
-            class="btn btn-group-sm w-100"
-            v-bind:class="activeSubmenu('Users')"
+            class="btn btn-group-sm w-100 inner-nav btn-secondary"
           >{{ $I18n.trans('users.users') }}</router-link>
           <router-link
             :to="usersCreateRoute"
-            class="btn btn-group-sm w-100"
-            v-bind:class="activeSubmenu('UsersCreate')"
+            class="btn btn-group-sm w-100 inner-nav btn-secondary"
           >{{ $I18n.trans('users.new-user') }}</router-link>
         </div>
       </div>
@@ -33,12 +31,6 @@ export default {
       usersRoute: UsersRoute,
       usersCreateRoute: UsersCreateRoute,
     };
-  },
-
-  methods: {
-    activeSubmenu(routeName) {
-      return this.$route.name === routeName ? 'btn-primary' : 'btn-secondary';
-    },
   },
 }
 </script>
